feat: break app.js into ES modules
Split public/app.js (3,405 lines) into 10 native ES modules in
public/js/ loaded via <script type="module">. No build step.

Modules: utils, state, theme, modal, backends, websocket, rendering,
conversations, streaming, main (entry point).

Shared mutable state centralized in state.js. Circular dependencies
avoided via late-binding callbacks. Inline onclick functions assigned
to window in main.js.

Also cleans stale SSE comments in src/routes/chat.ts.

Fixes #73
