--- public/app.js
@@ -1,3405 +1,0 @@
-// ─── HTML escape ─────────────────────────────────────────────────────────────
-function esc(str) {
-  return String(str || '')
-    .replace(/&/g, '&amp;')
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;')
-    .replace(/"/g, '&quot;');
-}
-
-function escWithCode(str) {
-  return esc(str).replace(/`([^`]+)`/g, '<code>$1</code>');
-}
-
-// ─── Timestamp / elapsed formatting ─────────────────────────────────────────
-function chatFormatTimestamp(isoString) {
-  if (!isoString) return '';
-  const date = new Date(isoString);
-  if (isNaN(date.getTime())) return '';
-  const now = new Date();
-  const isToday = date.toDateString() === now.toDateString();
-  const yesterday = new Date(now);
-  yesterday.setDate(yesterday.getDate() - 1);
-  const isYesterday = date.toDateString() === yesterday.toDateString();
-  const timeStr = date.toLocaleTimeString([], { hour: 'numeric', minute: '2-digit' });
-  if (isToday) return timeStr;
-  if (isYesterday) return `Yesterday ${timeStr}`;
-  return date.toLocaleDateString([], { month: 'short', day: 'numeric' }) + ' ' + timeStr;
-}
-
-function chatFormatElapsed(ms) {
-  const totalSec = Math.floor(ms / 1000);
-  const min = Math.floor(totalSec / 60);
-  const sec = totalSec % 60;
-  if (min === 0) return `${sec}s`;
-  return `${min}m ${sec < 10 ? '0' : ''}${sec}s`;
-}
-
-function chatFormatElapsedShort(ms) {
-  if (ms < 10000) return (ms / 1000).toFixed(1) + 's';
-  return chatFormatElapsed(ms);
-}
-
-// ─── Theme ───────────────────────────────────────────────────────────────────
-function applyTheme(theme) {
-  let resolved = theme;
-  if (theme === 'system') {
-    resolved = window.matchMedia('(prefers-color-scheme: dark)').matches ? 'dark' : 'light';
-  }
-  document.documentElement.setAttribute('data-theme', resolved);
-  const hljsLink = document.getElementById('hljs-theme');
-  if (hljsLink) {
-    const hljsStyle = resolved === 'dark' ? 'github-dark' : 'github';
-    hljsLink.href = `https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.9.0/styles/${hljsStyle}.min.css`;
-  }
-  try { localStorage.setItem('agent-cockpit-theme', theme); } catch (e) {}
-}
-
-window.matchMedia('(prefers-color-scheme: dark)').addEventListener('change', () => {
-  const cached = localStorage.getItem('agent-cockpit-theme') || 'system';
-  if (cached === 'system') applyTheme('system');
-});
-
-applyTheme(localStorage.getItem('agent-cockpit-theme') || 'system');
-
-// ─── State ────────────────────────────────────────────────────────────────────
-let csrfToken = null;
-
-const API_BASE = new URL('./api/', window.location.href);
-function apiUrl(path = '') {
-  const clean = String(path || '').replace(/^\/+/, '');
-  return new URL(clean, API_BASE).toString();
-}
-
-// ─── CSRF ─────────────────────────────────────────────────────────────────────
-async function fetchCsrfToken() {
-  const res = await fetch(apiUrl('/csrf-token'), { credentials: 'same-origin' });
-  if (!res.ok) throw new Error(`CSRF token fetch failed (${res.status})`);
-  const body = await res.json();
-  csrfToken = body.csrfToken;
-}
-
-// ─── Chat UI ──────────────────────────────────────────────────────────────────
-
-let CHAT_BACKENDS = [];
-let BACKEND_CAPABILITIES = {};
-let BACKEND_ICONS = {};
-const DEFAULT_BACKEND_ICON = '<svg width="28" height="28" viewBox="0 0 512 512" fill="none" xmlns="http://www.w3.org/2000/svg"><rect width="512" height="512" rx="128" fill="#888"/><text x="256" y="320" text-anchor="middle" fill="#fff" font-size="280" font-family="sans-serif">⚡</text></svg>';
-
-async function loadBackends() {
-  try {
-    const res = await fetch(apiUrl('/chat/backends'), { credentials: 'same-origin' });
-    if (!res.ok) throw new Error(`Failed to load backends (${res.status})`);
-    const data = await res.json();
-    CHAT_BACKENDS = data.backends.map(b => ({ id: b.id, label: b.label }));
-    for (const b of data.backends) {
-      BACKEND_CAPABILITIES[b.id] = b.capabilities || {};
-      BACKEND_ICONS[b.id] = b.icon || null;
-    }
-    populateBackendSelects();
-  } catch (err) {
-    console.error('[loadBackends]', err);
-    // Fallback so the UI still works
-    CHAT_BACKENDS = [{ id: 'claude-code', label: 'Claude Code' }];
-    populateBackendSelects();
-  }
-}
-
-function populateBackendSelects() {
-  const selects = document.querySelectorAll('#chat-backend-select, #chat-settings-backend');
-  for (const sel of selects) {
-    const current = sel.value;
-    sel.innerHTML = CHAT_BACKENDS.map(b =>
-      `<option value="${esc(b.id)}">${esc(b.label)}</option>`
-    ).join('');
-    if (current && [...sel.options].some(o => o.value === current)) {
-      sel.value = current;
-    }
-  }
-}
-
-function getBackendIcon(backendId) {
-  return BACKEND_ICONS[backendId] || DEFAULT_BACKEND_ICON;
-}
-
-function getBackendCapabilities(backendId) {
-  return BACKEND_CAPABILITIES[backendId] || {};
-}
-
-// Chat state
-let chatConversations = [];
-let chatActiveConvId = null;
-let chatActiveConv = null;
-let chatStreamingConvs = new Set();
-let chatResettingConvs = new Set();
-let chatStreamingState = new Map(); // convId -> { assistantContent, assistantThinking, activeTools, activeAgents, planModeActive, pendingInteraction, streamingMsgEl }
-let chatAbortController = null;
-let chatSidebarCollapsed = false;
-let chatSearchTimeout = null;
-let chatContextMenuEl = null;
-let chatSettingsData = null;
-let chatInitialized = false;
-let chatPendingWorkingDir = null;
-let chatPendingFiles = []; // Each: { file, status: 'uploading'|'done'|'error', progress, result, xhr }
-let chatDraftState = new Map(); // convId|'__new__' -> { text, pendingFiles }
-let _ensureConvPromise = null;
-let chatConvLoadGen = 0; // generation counter for chatLoadConversations to discard stale responses
-let chatMessageQueue = new Map(); // convId -> [{ id, content, inFlight }]
-let chatQueuePaused = new Set(); // convIds where queue is paused due to error
-let chatQueueIdCounter = 0;
-let chatWebSockets = new Map(); // convId -> WebSocket
-let chatReconnectAttempts = new Map(); // convId -> attempt count
-const CHAT_MAX_RECONNECT_ATTEMPTS = 5;
-const CHAT_RECONNECT_BASE_DELAY = 1000; // 1s, doubles each attempt
-
-// ── WebSocket helpers ───────────────────────────────────────────────────────
-
-function chatWsUrl(convId) {
-  const protocol = location.protocol === 'https:' ? 'wss:' : 'ws:';
-  return `${protocol}//${location.host}/api/chat/conversations/${convId}/ws`;
-}
-
-function chatConnectWs(convId) {
-  const existing = chatWebSockets.get(convId);
-  if (existing && (existing.readyState === WebSocket.OPEN || existing.readyState === WebSocket.CONNECTING)) {
-    return existing;
-  }
-  const ws = new WebSocket(chatWsUrl(convId));
-  chatWebSockets.set(convId, ws);
-
-  ws.onopen = () => {
-    console.log(`[ws] Connected for conv=${convId}`);
-  };
-
-  ws.onmessage = (evt) => {
-    try {
-      const event = JSON.parse(evt.data);
-      chatHandleStreamEvent(convId, event);
-    } catch (err) {
-      console.warn('[ws] Parse error:', err);
-    }
-  };
-
-  ws.onclose = (evt) => {
-    console.log(`[ws] Closed for conv=${convId}: code=${evt.code}`);
-    if (chatWebSockets.get(convId) === ws) {
-      chatWebSockets.delete(convId);
-    }
-    // Auto-reconnect if stream is still active (unexpected disconnect)
-    if (chatStreamingConvs.has(convId)) {
-      chatAutoReconnectWs(convId);
-    }
-  };
-
-  ws.onerror = (err) => {
-    console.error(`[ws] Error for conv=${convId}:`, err);
-  };
-
-  return ws;
-}
-
-function chatAutoReconnectWs(convId) {
-  const attempts = chatReconnectAttempts.get(convId) || 0;
-  if (attempts >= CHAT_MAX_RECONNECT_ATTEMPTS) {
-    console.warn(`[ws] Reconnect attempts exhausted for conv=${convId}`);
-    chatReconnectAttempts.delete(convId);
-    // Give up — resolve done so the stream cleans up
-    const st = chatStreamingState.get(convId);
-    if (st && st._doneResolve) { st._doneResolve(); delete st._doneResolve; }
-    return;
-  }
-  const delay = CHAT_RECONNECT_BASE_DELAY * Math.pow(2, attempts);
-  chatReconnectAttempts.set(convId, attempts + 1);
-  console.log(`[ws] Reconnecting conv=${convId} in ${delay}ms (attempt ${attempts + 1}/${CHAT_MAX_RECONNECT_ATTEMPTS})`);
-  setTimeout(() => {
-    // Stream may have ended while we waited
-    if (!chatStreamingConvs.has(convId)) {
-      chatReconnectAttempts.delete(convId);
-      return;
-    }
-    chatConnectWs(convId);
-  }, delay);
-}
-
-function chatDisconnectWs(convId) {
-  chatReconnectAttempts.delete(convId); // Prevent auto-reconnect on deliberate close
-  const ws = chatWebSockets.get(convId);
-  if (ws) {
-    ws.close();
-    chatWebSockets.delete(convId);
-  }
-}
-
-function chatWsSend(convId, frame) {
-  const ws = chatWebSockets.get(convId);
-  if (ws && ws.readyState === WebSocket.OPEN) {
-    ws.send(JSON.stringify(frame));
-    return true;
-  }
-  return false;
-}
-
-function chatApiUrl(path) {
-  return apiUrl('chat/' + path);
-}
-
-async function chatFetch(path, opts = {}) {
-  if (!csrfToken) await fetchCsrfToken();
-  const headers = { ...opts.headers };
-  if (csrfToken) headers['x-csrf-token'] = csrfToken;
-  if (opts.body && typeof opts.body === 'object' && !(opts.body instanceof FormData)) {
-    headers['Content-Type'] = 'application/json';
-    opts.body = JSON.stringify(opts.body);
-  }
-  const res = await fetch(chatApiUrl(path), { ...opts, headers, credentials: 'same-origin' });
-  if (!res.ok) {
-    const err = await res.json().catch(() => ({ error: res.statusText }));
-    throw new Error(err.error || res.statusText);
-  }
-  return res;
-}
-
-function chatInit() {
-  if (chatInitialized && chatActiveConvId) {
-    // Re-entering — just refresh
-    chatRenderConvList();
-    if (chatActiveConv) chatRenderMessages();
-    chatWireEvents();
-    return;
-  }
-  chatInitialized = true;
-  chatWireEvents();
-  loadBackends();
-  chatLoadConversations();
-
-  // Show app version in sidebar + check for updates
-  chatFetch('version').then(res => res.json()).then(v => {
-    const textEl = document.getElementById('chat-version-text');
-    if (textEl && v.version) textEl.textContent = 'v' + v.version;
-    chatCheckUpdateIndicator(v);
-  }).catch(() => {});
-
-  // Click version number to trigger a manual version check
-  const versionLabel = document.getElementById('chat-version-text');
-  if (versionLabel) {
-    versionLabel.style.cursor = 'pointer';
-    versionLabel.title = 'Click to check for updates';
-    versionLabel.addEventListener('click', chatManualVersionCheck);
-  }
-
-  // Poll for update status every 5 minutes (reads server-cached state, no git ops)
-  setInterval(() => {
-    chatFetch('update-status').then(res => res.json()).then(chatCheckUpdateIndicator).catch(() => {});
-  }, 5 * 60 * 1000);
-
-  // Sync theme from server settings
-  chatFetch('settings').then(res => res.json()).then(s => {
-    chatSettingsData = s;
-    applyTheme(s.theme || 'system');
-  }).catch(() => {});
-}
-
-function chatWireEvents() {
-  // Use event delegation for conversation list — attaching handlers to each
-  // child after every innerHTML rebuild causes lost clicks when the DOM is
-  // rebuilt between mousedown and mouseup (e.g. during streaming re-renders).
-  const convList = document.getElementById('chat-conv-list');
-  if (convList && !convList._delegated) {
-    convList._delegated = true;
-    convList.addEventListener('click', (e) => {
-      // Workspace instructions button
-      const instrBtn = e.target.closest('.chat-conv-group-instructions-btn');
-      if (instrBtn) {
-        e.stopPropagation();
-        chatShowWorkspaceInstructions(instrBtn.dataset.wsHash, instrBtn.dataset.wsLabel);
-        return;
-      }
-      // Group header toggle
-      const groupHeader = e.target.closest('.chat-conv-group-header');
-      if (groupHeader) {
-        const grp = groupHeader.dataset.group;
-        const isNowCollapsed = !chatGetCollapsedGroups()[grp];
-        chatSetGroupCollapsed(grp, isNowCollapsed);
-        chatRenderConvList();
-        return;
-      }
-      // Context menu button
-      const menuBtn = e.target.closest('.chat-conv-item-menu');
-      if (menuBtn) {
-        e.stopPropagation();
-        chatShowContextMenu(e, menuBtn.dataset.convMenu);
-        return;
-      }
-      // Conversation item selection
-      const convItem = e.target.closest('.chat-conv-item');
-      if (convItem) {
-        chatSelectConversation(convItem.dataset.convId);
-        return;
-      }
-    });
-  }
-
-  const newBtn = document.getElementById('chat-new-btn');
-  if (newBtn) newBtn.onclick = chatNewConversation;
-
-  const collapseBtn = document.getElementById('chat-sidebar-collapse');
-  if (collapseBtn) collapseBtn.onclick = () => chatToggleSidebar();
-
-  const toggleBtn = document.getElementById('chat-header-toggle');
-  if (toggleBtn) toggleBtn.onclick = () => chatToggleSidebar();
-
-  const sendBtn = document.getElementById('chat-send-btn');
-  if (sendBtn) sendBtn.onclick = () => {
-    const ta = document.getElementById('chat-textarea');
-    const hasInput = ta && ta.value.trim();
-    if (chatStreamingConvs.has(chatActiveConvId) && !hasInput) chatStopStreaming();
-    else chatSendMessage();
-  };
-
-  const textarea = document.getElementById('chat-textarea');
-  if (textarea) {
-    textarea.oninput = () => {
-      chatAutoResize(textarea);
-      chatUpdateSendButtonState();
-    };
-    textarea.onkeydown = (e) => {
-      if (e.key === 'Enter' && !e.shiftKey) {
-        e.preventDefault();
-        const hasUploading = chatPendingFiles.some(entry => entry.status === 'uploading');
-        if (!hasUploading && (textarea.value.trim() || chatPendingFiles.some(entry => entry.status === 'done'))) {
-          chatSendMessage();
-        }
-      }
-    };
-    textarea.addEventListener('paste', (e) => {
-      const items = e.clipboardData?.items;
-      if (!items) return;
-      const files = [];
-      for (const item of items) {
-        if (item.kind === 'file') {
-          const file = item.getAsFile();
-          if (file) {
-            const ts = Date.now();
-            const ext = file.name.includes('.') ? '.' + file.name.split('.').pop() : '.png';
-            const baseName = file.name.replace(/\.[^.]+$/, '');
-            const uniqueName = `${baseName}-${ts}-${files.length + 1}${ext}`;
-            const renamed = new File([file], uniqueName, { type: file.type });
-            files.push(renamed);
-          }
-        }
-      }
-      if (files.length) {
-        e.preventDefault();
-        chatAddPendingFiles(files);
-        return;
-      }
-      // Convert large text pastes (1000+ chars) to text file attachments
-      const pastedText = e.clipboardData.getData('text/plain');
-      if (pastedText && pastedText.length >= 1000) {
-        e.preventDefault();
-        const now = new Date();
-        const ts = now.getFullYear()
-          + String(now.getMonth() + 1).padStart(2, '0')
-          + String(now.getDate()).padStart(2, '0')
-          + '-'
-          + String(now.getHours()).padStart(2, '0')
-          + String(now.getMinutes()).padStart(2, '0')
-          + String(now.getSeconds()).padStart(2, '0');
-        const textFile = new File([pastedText], `pasted-text-${ts}.txt`, { type: 'text/plain' });
-        chatAddPendingFiles([textFile]);
-      }
-    });
-  }
-
-  const searchInput = document.getElementById('chat-search-input');
-  if (searchInput) {
-    searchInput.oninput = () => {
-      clearTimeout(chatSearchTimeout);
-      chatSearchTimeout = setTimeout(() => chatLoadConversations(searchInput.value), 300);
-    };
-  }
-
-  const downloadBtn = document.getElementById('chat-download-btn');
-  if (downloadBtn) downloadBtn.onclick = chatDownloadConversation;
-
-  const resetBtn = document.getElementById('chat-reset-btn');
-  if (resetBtn) resetBtn.onclick = chatResetSession;
-
-  const sessionsBtn = document.getElementById('chat-sessions-btn');
-  if (sessionsBtn) sessionsBtn.onclick = chatShowSessions;
-
-  const settingsBtn = document.getElementById('chat-settings-btn');
-  if (settingsBtn) settingsBtn.onclick = chatShowSettings;
-
-  const signoutBtn = document.getElementById('chat-signout-btn');
-  if (signoutBtn) signoutBtn.onclick = () => {
-    if (confirm('Sign out?')) window.location.href = '/auth/logout';
-  };
-
-  // Prompt cards
-  document.querySelectorAll('.chat-prompt-card').forEach(card => {
-    card.onclick = () => {
-      const textarea = document.getElementById('chat-textarea');
-      if (textarea) {
-        textarea.value = card.dataset.prompt;
-        chatAutoResize(textarea);
-        chatSendMessage();
-      }
-    };
-  });
-
-  // ── File attachment wiring ──
-  const attachBtn = document.getElementById('chat-attach-btn');
-  const fileInput = document.getElementById('chat-file-input');
-  if (attachBtn && fileInput) {
-    attachBtn.onclick = () => fileInput.click();
-    fileInput.onchange = () => {
-      if (fileInput.files.length) {
-        chatAddPendingFiles(Array.from(fileInput.files));
-        fileInput.value = '';
-      }
-    };
-  }
-
-  // Drag-and-drop on chat messages area
-  const chatMsgs = document.getElementById('chat-messages');
-  if (chatMsgs) {
-    let dragCounter = 0;
-    chatMsgs.addEventListener('dragenter', (e) => {
-      e.preventDefault();
-      dragCounter++;
-      if (dragCounter === 1) chatShowDropOverlay(true);
-    });
-    chatMsgs.addEventListener('dragover', (e) => e.preventDefault());
-    chatMsgs.addEventListener('dragleave', (e) => {
-      e.preventDefault();
-      dragCounter--;
-      if (dragCounter <= 0) { dragCounter = 0; chatShowDropOverlay(false); }
-    });
-    chatMsgs.addEventListener('drop', (e) => {
-      e.preventDefault();
-      dragCounter = 0;
-      chatShowDropOverlay(false);
-      if (e.dataTransfer.files.length) chatAddPendingFiles(Array.from(e.dataTransfer.files));
-    });
-  }
-
-  // Close context menu on click elsewhere
-  document.addEventListener('click', chatCloseContextMenu);
-}
-
-function chatAutoResize(el) {
-  el.style.height = 'auto';
-  el.style.height = Math.min(el.scrollHeight, 200) + 'px';
-}
-
-// ── File attachment helpers ──────────────────────────────────────────────────
-
-function chatFormatFileSize(bytes) {
-  if (bytes < 1024) return bytes + ' B';
-  if (bytes < 1024 * 1024) return (bytes / 1024).toFixed(1) + ' KB';
-  return (bytes / (1024 * 1024)).toFixed(1) + ' MB';
-}
-
-async function chatEnsureConversation() {
-  if (chatActiveConvId) return chatActiveConvId;
-  if (_ensureConvPromise) return _ensureConvPromise;
-  _ensureConvPromise = (async () => {
-    try {
-      const body = chatPendingWorkingDir ? { workingDir: chatPendingWorkingDir } : {};
-      chatPendingWorkingDir = null;
-      const res = await chatFetch('conversations', { method: 'POST', body });
-      const conv = await res.json();
-      // Migrate __new__ draft to real conversation ID
-      if (chatDraftState.has('__new__')) {
-        chatDraftState.set(conv.id, chatDraftState.get('__new__'));
-        chatDraftState.delete('__new__');
-      }
-      chatActiveConvId = conv.id;
-      chatActiveConv = conv;
-      chatLoadConversations();
-      chatUpdateHeader();
-      chatRenderMessages();
-      return conv.id;
-    } finally {
-      _ensureConvPromise = null;
-    }
-  })();
-  return _ensureConvPromise;
-}
-
-async function chatUploadSingleFile(convId, entry) {
-  if (!csrfToken) await fetchCsrfToken();
-  return new Promise((resolve) => {
-    const xhr = new XMLHttpRequest();
-    entry.xhr = xhr;
-    const formData = new FormData();
-    formData.append('files', entry.file);
-    xhr.open('POST', chatApiUrl(`conversations/${convId}/upload`));
-    xhr.setRequestHeader('x-csrf-token', csrfToken || '');
-    xhr.withCredentials = true;
-    xhr.upload.onprogress = (e) => {
-      if (e.lengthComputable) {
-        entry.progress = Math.round((e.loaded / e.total) * 100);
-        chatRenderFileChips();
-      }
-    };
-    xhr.onload = () => {
-      entry.xhr = null;
-      if (xhr.status >= 200 && xhr.status < 300) {
-        try {
-          const resp = JSON.parse(xhr.responseText);
-          entry.status = 'done';
-          entry.result = resp.files[0];
-        } catch {
-          entry.status = 'error';
-        }
-      } else {
-        entry.status = 'error';
-      }
-      chatRenderFileChips();
-      chatUpdateSendButtonState();
-      resolve();
-    };
-    xhr.onerror = () => {
-      entry.xhr = null;
-      entry.status = 'error';
-      chatRenderFileChips();
-      chatUpdateSendButtonState();
-      resolve();
-    };
-    xhr.onabort = () => {
-      entry.xhr = null;
-      resolve();
-    };
-    xhr.send(formData);
-  });
-}
-
-async function chatAddPendingFiles(files) {
-  const newEntries = files.map(f => ({ file: f, status: 'uploading', progress: 0, result: null, xhr: null }));
-  chatPendingFiles.push(...newEntries);
-  chatRenderFileChips();
-  chatUpdateSendButtonState();
-  try {
-    const convId = await chatEnsureConversation();
-    for (const entry of newEntries) {
-      if (entry.status === 'uploading') {
-        chatUploadSingleFile(convId, entry);
-      }
-    }
-  } catch (err) {
-    for (const entry of newEntries) {
-      entry.status = 'error';
-      entry.xhr = null;
-    }
-    chatRenderFileChips();
-    chatUpdateSendButtonState();
-    alert('Failed to create conversation: ' + err.message);
-  }
-}
-
-function chatRemovePendingFile(index) {
-  const entry = chatPendingFiles[index];
-  if (!entry) return;
-  if (entry.status === 'uploading' && entry.xhr) {
-    entry.xhr.abort();
-  }
-  if (entry.status === 'done' && entry.result && chatActiveConvId) {
-    chatFetch(`conversations/${chatActiveConvId}/upload/${encodeURIComponent(entry.result.name)}`, { method: 'DELETE' }).catch(() => {});
-  }
-  chatPendingFiles.splice(index, 1);
-  chatRenderFileChips();
-  chatUpdateSendButtonState();
-}
-
-function chatRenderFileChips() {
-  const container = document.getElementById('chat-file-chips');
-  if (!container) return;
-  if (!chatPendingFiles.length) {
-    container.style.display = 'none';
-    container.innerHTML = '';
-    return;
-  }
-  container.style.display = 'flex';
-  container.innerHTML = chatPendingFiles.map((entry, i) => {
-    const f = entry.file;
-    const isImage = f.type && f.type.startsWith('image/');
-    const thumbHtml = isImage ? `<img class="chat-file-chip-thumb" src="${URL.createObjectURL(f)}" alt="">` : '';
-    let statusHtml = '';
-    if (entry.status === 'uploading') {
-      statusHtml = `<div class="chat-file-chip-progress"><div class="chat-file-chip-progress-bar" style="width:${entry.progress}%"></div></div>`;
-    }
-    const doneIcon = entry.status === 'done' ? '<span class="chat-file-chip-done">&#10003;</span>' : '';
-    const errorIcon = entry.status === 'error' ? '<span class="chat-file-chip-error" title="Upload failed">!</span>' : '';
-    return `<div class="chat-file-chip${entry.status === 'error' ? ' error' : ''}">
-      ${thumbHtml}
-      <span class="chat-file-chip-name" title="${esc(f.name)}">${esc(f.name)}</span>
-      <span class="chat-file-chip-size">${chatFormatFileSize(f.size)}</span>
-      ${doneIcon}${errorIcon}
-      ${statusHtml}
-      <button class="chat-file-chip-remove" data-file-index="${i}" title="Remove">&times;</button>
-    </div>`;
-  }).join('');
-  container.querySelectorAll('.chat-file-chip-remove').forEach(btn => {
-    btn.addEventListener('click', () => {
-      chatRemovePendingFile(Number(btn.dataset.fileIndex));
-    });
-  });
-}
-
-function chatShowDropOverlay(show) {
-  const chatMsgs = document.getElementById('chat-messages');
-  if (!chatMsgs) return;
-  let overlay = chatMsgs.querySelector('.chat-drop-overlay');
-  if (show && !overlay) {
-    chatMsgs.style.position = 'relative';
-    overlay = document.createElement('div');
-    overlay.className = 'chat-drop-overlay';
-    overlay.textContent = 'Drop files here';
-    chatMsgs.appendChild(overlay);
-  } else if (!show && overlay) {
-    overlay.remove();
-  }
-}
-
-async function chatUploadFiles(convId, files) {
-  if (!csrfToken) await fetchCsrfToken();
-  const formData = new FormData();
-  files.forEach(f => formData.append('files', f));
-  const res = await fetch(chatApiUrl(`conversations/${convId}/upload`), {
-    method: 'POST',
-    headers: { 'x-csrf-token': csrfToken || '' },
-    credentials: 'same-origin',
-    body: formData,
-  });
-  if (!res.ok) {
-    const err = await res.json().catch(() => ({}));
-    throw new Error(err.error || 'Upload failed');
-  }
-  return res.json();
-}
-
-function chatToggleSidebar() {
-  chatSidebarCollapsed = !chatSidebarCollapsed;
-  const sidebar = document.getElementById('chat-sidebar');
-  if (sidebar) sidebar.classList.toggle('collapsed', chatSidebarCollapsed);
-}
-
-// ── Conversation list ─────────────────────────────────────────────────────────
-
-async function chatLoadConversations(query) {
-  const gen = ++chatConvLoadGen;
-  try {
-    const q = query ? `?q=${encodeURIComponent(query)}` : '';
-    const res = await chatFetch(`conversations${q}`);
-    if (gen !== chatConvLoadGen) return; // stale response — a newer call was made
-    const data = await res.json();
-    chatConversations = data.conversations || [];
-    chatRenderConvList();
-    // Sync active conversation title with authoritative server data
-    if (chatActiveConv && chatActiveConvId) {
-      const match = chatConversations.find(c => c.id === chatActiveConvId);
-      if (match && match.title !== chatActiveConv.title) {
-        chatActiveConv.title = match.title;
-        chatUpdateHeader();
-      }
-    }
-  } catch (err) {
-    console.error('Failed to load conversations:', err);
-  }
-}
-
-function chatGroupConversations(convs) {
-  const groups = {};
-  for (const c of convs) {
-    const label = c.workingDir
-      ? c.workingDir.split('/').filter(Boolean).slice(-2).join('/')
-      : 'workspace';
-    if (!groups[label]) groups[label] = { fullPath: c.workingDir || '', hash: c.workspaceHash || '', convs: [] };
-    groups[label].convs.push(c);
-  }
-  return groups;
-}
-
-function chatGetCollapsedGroups() {
-  try {
-    return JSON.parse(localStorage.getItem('chatCollapsedGroups') || '{}');
-  } catch { return {}; }
-}
-
-function chatSetGroupCollapsed(label, collapsed) {
-  const state = chatGetCollapsedGroups();
-  if (collapsed) state[label] = true;
-  else delete state[label];
-  localStorage.setItem('chatCollapsedGroups', JSON.stringify(state));
-}
-
-function chatRenderConvList() {
-  const list = document.getElementById('chat-conv-list');
-  if (!list) return;
-
-  if (chatConversations.length === 0) {
-    list.innerHTML = '<div style="padding:20px;text-align:center;color:var(--muted);font-size:12px;">No conversations yet</div>';
-    return;
-  }
-
-  const groups = chatGroupConversations(chatConversations);
-  const collapsed = chatGetCollapsedGroups();
-  let html = '';
-  for (const [label, group] of Object.entries(groups)) {
-    const isCollapsed = !!collapsed[label];
-    const count = group.convs.length;
-    html += `
-      <div class="chat-conv-group-header" data-group="${esc(label)}" title="${esc(group.fullPath || 'Default workspace')}">
-        <svg class="chat-conv-group-chevron${isCollapsed ? ' collapsed' : ''}" width="16" height="16" viewBox="0 0 16 16" fill="none"><path d="M4.5 6L8 9.5L11.5 6" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"/></svg>
-        <span class="chat-conv-group-label">${esc(label)}</span>
-        ${isCollapsed ? `<span class="chat-conv-group-count">${count}</span>` : ''}
-        ${group.hash ? `<button class="chat-conv-group-instructions-btn" data-ws-hash="${esc(group.hash)}" data-ws-label="${esc(label)}" title="Workspace instructions"><svg width="14" height="14" viewBox="0 0 16 16" fill="none"><path d="M13.5 4.5l-2-2L3 11l-.5 2.5L5 13l8.5-8.5z" stroke="currentColor" stroke-width="1.3" stroke-linecap="round" stroke-linejoin="round"/><path d="M10.5 3.5l2 2" stroke="currentColor" stroke-width="1.3" stroke-linecap="round"/></svg></button>` : ''}
-      </div>`;
-    if (!isCollapsed) {
-      for (const c of group.convs) {
-        const isActive = c.id === chatActiveConvId;
-        const isStreaming = chatStreamingConvs.has(c.id);
-        html += `
-          <div class="chat-conv-item${isActive ? ' active' : ''}" data-conv-id="${esc(c.id)}">
-            <div style="flex:1;min-width:0;">
-              <span class="chat-conv-item-title">${esc(c.title)}</span>
-            </div>
-            ${isStreaming ? '<span class="chat-conv-streaming-dot"></span>' : ''}
-            <button class="chat-conv-item-menu" data-conv-menu="${esc(c.id)}" title="Options">⋯</button>
-          </div>
-        `;
-      }
-    }
-  }
-  list.innerHTML = html;
-}
-
-// ── Conversation operations ───────────────────────────────────────────────────
-
-async function chatNewConversation() {
-  chatShowFolderPicker();
-}
-
-async function chatCreateConversationWithDir(workingDir) {
-  try {
-    // Save draft from current conversation before switching
-    chatSaveDraft();
-    for (const entry of chatPendingFiles) {
-      if (entry.status === 'uploading' && entry.xhr) entry.xhr.abort();
-    }
-    const body = workingDir ? { workingDir } : {};
-    const res = await chatFetch('conversations', { method: 'POST', body });
-    const conv = await res.json();
-    chatActiveConvId = conv.id;
-    chatActiveConv = conv;
-    await chatLoadConversations();
-    chatRenderMessages();
-    chatUpdateHeader();
-    chatRestoreDraft(conv.id);
-    chatCloseModal();
-    const textarea = document.getElementById('chat-textarea');
-    if (textarea) textarea.focus();
-  } catch (err) {
-    alert('Failed to create conversation: ' + err.message);
-  }
-}
-
-async function chatShowFolderPicker(initialPath) {
-  const browsePath = initialPath || '';
-  let showHidden = false;
-
-  async function loadDir(dirPath) {
-    try {
-      const q = dirPath ? `?path=${encodeURIComponent(dirPath)}` : '';
-      const showQ = showHidden ? (q ? '&showHidden=true' : '?showHidden=true') : '';
-      const res = await chatFetch(`browse${q}${showQ}`);
-      const data = await res.json();
-      renderFolderBrowser(data);
-    } catch (err) {
-      const pathEl = document.querySelector('.folder-browser-path');
-      if (pathEl) {
-        const origText = pathEl.textContent;
-        pathEl.textContent = '⚠️ ' + (err.message || 'Cannot access folder');
-        pathEl.style.color = 'var(--blocked, #dc2626)';
-        setTimeout(() => { pathEl.textContent = origText; pathEl.style.color = ''; }, 2000);
-      }
-    }
-  }
-
-  function renderFolderBrowser(data) {
-    const folderName = data.currentPath.split('/').pop() || data.currentPath;
-    let listHtml = '';
-    if (data.parent) {
-      listHtml += `<div class="folder-browser-item parent-item" data-path="${esc(data.parent)}">↑ Parent Directory</div>`;
-    }
-    if (data.dirs.length === 0) {
-      listHtml += '<div style="padding:12px;color:var(--muted);font-size:12px;text-align:center;">No subdirectories</div>';
-    }
-    for (const d of data.dirs) {
-      const full = data.currentPath + '/' + d;
-      listHtml += `<div class="folder-browser-item" data-path="${esc(full)}">📁 ${esc(d)}</div>`;
-    }
-
-    const bodyHtml = `
-      <div class="chat-modal-body">
-        <div class="folder-browser-path-row">
-          <div class="folder-browser-path" title="${esc(data.currentPath)}">${esc(data.currentPath)}</div>
-          <div class="folder-browser-path-actions">
-            ${data.parent ? `<button class="folder-browser-icon-btn" id="folder-go-parent" title="Go to parent folder">&#x2191;</button>` : ''}
-            ${data.parent ? `<button class="folder-browser-icon-btn folder-browser-delete-btn" id="folder-delete-btn" title="Delete this folder">&#x1F5D1;</button>` : ''}
-          </div>
-        </div>
-        <div class="folder-browser-toolbar">
-          <label class="folder-browser-toggle">
-            <input type="checkbox" id="folder-show-hidden" ${showHidden ? 'checked' : ''} /> Show hidden folders
-          </label>
-          <button class="folder-browser-new-btn" id="folder-new-btn" title="New Folder">+ New Folder</button>
-        </div>
-        <div class="folder-browser-new-input" id="folder-new-input" style="display:none;">
-          <input type="text" id="folder-new-name" placeholder="Folder name" autocomplete="off" />
-          <button id="folder-new-confirm" title="Create">&#10003;</button>
-          <button id="folder-new-cancel" title="Cancel">&#10005;</button>
-        </div>
-        <div class="folder-browser-list" id="folder-browser-list">${listHtml}</div>
-      </div>
-      <div class="folder-browser-actions" style="padding:12px 20px;border-top:1px solid var(--border);">
-        <button class="folder-browser-default" id="folder-use-default">Use Default (workspace)</button>
-        <button class="folder-browser-select" id="folder-select-this">Select This Folder</button>
-      </div>
-    `;
-
-    chatShowModal('Select Working Directory', bodyHtml);
-
-    // Wire events
-    document.getElementById('folder-browser-list').querySelectorAll('.folder-browser-item').forEach(el => {
-      el.addEventListener('click', () => loadDir(el.dataset.path));
-    });
-    document.getElementById('folder-show-hidden').onchange = (e) => {
-      showHidden = e.target.checked;
-      loadDir(data.currentPath);
-    };
-    document.getElementById('folder-select-this').onclick = () => chatCreateConversationWithDir(data.currentPath);
-    document.getElementById('folder-use-default').onclick = () => chatCreateConversationWithDir(null);
-
-    // New Folder logic
-    const newBtn = document.getElementById('folder-new-btn');
-    const newInputRow = document.getElementById('folder-new-input');
-    const newNameInput = document.getElementById('folder-new-name');
-    newBtn.onclick = () => {
-      newInputRow.style.display = 'flex';
-      newBtn.style.display = 'none';
-      newNameInput.value = '';
-      newNameInput.focus();
-    };
-    document.getElementById('folder-new-cancel').onclick = () => {
-      newInputRow.style.display = 'none';
-      newBtn.style.display = '';
-    };
-    async function createNewFolder() {
-      const name = newNameInput.value.trim();
-      if (!name) return;
-      try {
-        const res = await chatFetch('mkdir', { method: 'POST', body: { parentPath: data.currentPath, name } });
-        const result = await res.json();
-        loadDir(result.created);
-      } catch (err) {
-        const pathEl = document.querySelector('.folder-browser-path');
-        if (pathEl) {
-          const origText = pathEl.textContent;
-          const origColor = pathEl.style.color;
-          pathEl.textContent = '\u26a0\ufe0f ' + (err.message || 'Failed to create folder');
-          pathEl.style.color = 'var(--blocked, #dc2626)';
-          setTimeout(() => { pathEl.textContent = origText; pathEl.style.color = origColor; }, 2000);
-        }
-      }
-    }
-    document.getElementById('folder-new-confirm').onclick = createNewFolder;
-    newNameInput.onkeydown = (e) => {
-      if (e.key === 'Enter') createNewFolder();
-      if (e.key === 'Escape') { newInputRow.style.display = 'none'; newBtn.style.display = ''; }
-    };
-
-    // Parent navigation button
-    const goParentBtn = document.getElementById('folder-go-parent');
-    if (goParentBtn) goParentBtn.onclick = () => loadDir(data.parent);
-
-    // Delete folder button with confirmation
-    const deleteBtn = document.getElementById('folder-delete-btn');
-    if (deleteBtn) {
-      deleteBtn.onclick = () => {
-        const folderName = data.currentPath.split('/').pop() || data.currentPath;
-        const confirmEl = document.createElement('div');
-        confirmEl.className = 'folder-browser-confirm-delete';
-        confirmEl.innerHTML = `
-          <span>Delete <strong>${esc(folderName)}</strong> and all its contents?</span>
-          <div class="folder-browser-confirm-actions">
-            <button class="folder-browser-confirm-yes" id="folder-delete-yes">Delete</button>
-            <button class="folder-browser-confirm-no" id="folder-delete-no">Cancel</button>
-          </div>
-        `;
-        const list = document.getElementById('folder-browser-list');
-        if (list) list.replaceWith(confirmEl);
-        document.getElementById('folder-delete-no').onclick = () => renderFolderBrowser(data);
-        document.getElementById('folder-delete-yes').onclick = async () => {
-          try {
-            await chatFetch('rmdir', { method: 'POST', body: { dirPath: data.currentPath } });
-            loadDir(data.parent);
-          } catch (err) {
-            const pathEl = document.querySelector('.folder-browser-path');
-            if (pathEl) {
-              const origText = pathEl.textContent;
-              const origColor = pathEl.style.color;
-              pathEl.textContent = '\u26a0\ufe0f ' + (err.message || 'Failed to delete folder');
-              pathEl.style.color = 'var(--blocked, #dc2626)';
-              setTimeout(() => { pathEl.textContent = origText; pathEl.style.color = origColor; }, 2000);
-            }
-            renderFolderBrowser(data);
-          }
-        };
-      };
-    }
-  }
-
-  loadDir(browsePath);
-}
-window.chatShowFolderPicker = chatShowFolderPicker;
-
-function chatUpdateSendButtonState() {
-  const sendBtn = document.getElementById('chat-send-btn');
-  if (!sendBtn) return;
-  const isStreaming = chatStreamingConvs.has(chatActiveConvId);
-  const isResetting = chatResettingConvs.has(chatActiveConvId);
-  const ta = document.getElementById('chat-textarea');
-  const hasText = ta && ta.value.trim();
-  if (isStreaming) {
-    if (hasText) {
-      // Show send arrow to queue the message
-      sendBtn.disabled = false;
-      sendBtn.textContent = '↑';
-      sendBtn.classList.remove('stop');
-      sendBtn.title = 'Queue message (Enter)';
-    } else {
-      // Show stop button
-      sendBtn.disabled = false;
-      sendBtn.textContent = '■';
-      sendBtn.classList.add('stop');
-      sendBtn.title = 'Stop streaming';
-    }
-  } else {
-    sendBtn.textContent = '↑';
-    sendBtn.classList.remove('stop');
-    sendBtn.title = 'Send message (Enter)';
-    const hasCompletedFiles = chatPendingFiles.some(e => e.status === 'done');
-    const hasUploading = chatPendingFiles.some(e => e.status === 'uploading');
-    sendBtn.disabled = isResetting || hasUploading || (!hasText && !hasCompletedFiles);
-  }
-}
-
-function chatSaveDraft() {
-  const key = chatActiveConvId || '__new__';
-  const textarea = document.getElementById('chat-textarea');
-  const text = textarea ? textarea.value : '';
-  if (!text && !chatPendingFiles.length) {
-    chatDraftState.delete(key);
-    return;
-  }
-  chatDraftState.set(key, { text, pendingFiles: chatPendingFiles });
-}
-
-function chatRestoreDraft(convId) {
-  const key = convId || '__new__';
-  const draft = chatDraftState.get(key);
-  const textarea = document.getElementById('chat-textarea');
-  if (draft) {
-    if (textarea) {
-      textarea.value = draft.text;
-      chatAutoResize(textarea);
-    }
-    chatPendingFiles = draft.pendingFiles;
-  } else {
-    if (textarea) {
-      textarea.value = '';
-      chatAutoResize(textarea);
-    }
-    chatPendingFiles = [];
-  }
-  chatRenderFileChips();
-  chatUpdateSendButtonState();
-}
-
-async function chatSelectConversation(id) {
-  if (id === chatActiveConvId) return;
-  // Save current draft before switching
-  chatSaveDraft();
-  // Abort in-flight uploads from current conversation (they are saved as-is in draft)
-  for (const entry of chatPendingFiles) {
-    if (entry.status === 'uploading' && entry.xhr) entry.xhr.abort();
-  }
-  try {
-    const res = await chatFetch(`conversations/${id}`);
-    chatActiveConv = await res.json();
-    chatActiveConvId = id;
-    chatRenderConvList();
-    chatRenderMessages();
-    chatUpdateHeader();
-    chatRestoreDraft(id);
-    const resetBtn = document.getElementById('chat-reset-btn');
-    if (resetBtn) {
-      resetBtn.disabled = chatResettingConvs.has(id);
-      resetBtn.textContent = chatResettingConvs.has(id) ? '↻ Resetting...' : '↻ Reset';
-    }
-    const backendSelect = document.getElementById('chat-backend-select');
-    if (backendSelect && chatActiveConv.backend) {
-      backendSelect.value = chatActiveConv.backend;
-    }
-  } catch (err) {
-    alert('Failed to load conversation: ' + err.message);
-  }
-}
-
-async function chatRenameConversation(id) {
-  const conv = chatConversations.find(c => c.id === id);
-  const newTitle = prompt('Rename conversation:', conv ? conv.title : '');
-  if (!newTitle || !newTitle.trim()) return;
-  try {
-    await chatFetch(`conversations/${id}`, { method: 'PUT', body: { title: newTitle.trim() } });
-    if (chatActiveConvId === id && chatActiveConv) chatActiveConv.title = newTitle.trim();
-    chatUpdateHeader();
-    chatLoadConversations();
-  } catch (err) {
-    alert('Failed to rename: ' + err.message);
-  }
-}
-
-async function chatDeleteConversation(id) {
-  if (!confirm('Delete this conversation? This cannot be undone.')) return;
-  try {
-    await chatFetch(`conversations/${id}`, { method: 'DELETE' });
-    chatDraftState.delete(id);
-    if (chatActiveConvId === id) {
-      // Abort in-flight uploads and clear pending files
-      for (const entry of chatPendingFiles) {
-        if (entry.status === 'uploading' && entry.xhr) entry.xhr.abort();
-      }
-      chatPendingFiles = [];
-      chatRenderFileChips();
-      chatActiveConvId = null;
-      chatActiveConv = null;
-      chatRenderMessages();
-      chatUpdateHeader();
-      chatUpdateSendButtonState();
-    }
-    chatLoadConversations();
-  } catch (err) {
-    alert('Failed to delete: ' + err.message);
-  }
-}
-
-function chatUpdateHeader() {
-  const titleEl = document.getElementById('chat-header-title');
-  if (titleEl) {
-    titleEl.textContent = chatActiveConv ? chatActiveConv.title : 'New Chat';
-  }
-  let wdEl = document.getElementById('chat-header-workdir');
-  if (!wdEl) {
-    const header = titleEl?.parentElement;
-    if (header) {
-      wdEl = document.createElement('div');
-      wdEl.className = 'chat-header-workdir';
-      wdEl.id = 'chat-header-workdir';
-      titleEl.after(wdEl);
-    }
-  }
-  if (wdEl) {
-    if (chatActiveConv && chatActiveConv.workingDir) {
-      const folderName = chatActiveConv.workingDir.split('/').filter(Boolean).slice(-2).join('/') || chatActiveConv.workingDir;
-      wdEl.textContent = '📁 ' + folderName;
-      wdEl.title = chatActiveConv.workingDir;
-      wdEl.style.display = '';
-    } else if (chatActiveConv) {
-      wdEl.textContent = '📁 workspace';
-      wdEl.title = 'Default workspace';
-      wdEl.style.display = '';
-    } else {
-      wdEl.style.display = 'none';
-    }
-  }
-  chatUpdateUsageDisplay();
-}
-
-function chatFormatTokenCount(n) {
-  if (n >= 1000000) return (n / 1000000).toFixed(1) + 'M';
-  if (n >= 1000) return (n / 1000).toFixed(1) + 'k';
-  return String(n);
-}
-
-function chatFormatCost(usd) {
-  if (usd < 0.01) return '$' + usd.toFixed(4);
-  if (usd < 1) return '$' + usd.toFixed(3);
-  return '$' + usd.toFixed(2);
-}
-
-function chatUpdateUsageDisplay() {
-  let el = document.getElementById('chat-header-usage');
-  const convUsage = chatActiveConv?.usage;
-  const sessUsage = chatActiveConv?.sessionUsage;
-
-  // Show badge if either session or conversation has usage
-  const hasUsage = (sessUsage && (sessUsage.inputTokens > 0 || sessUsage.outputTokens > 0 || sessUsage.costUsd > 0))
-    || (convUsage && (convUsage.inputTokens > 0 || convUsage.outputTokens > 0 || convUsage.costUsd > 0));
-
-  if (!hasUsage) {
-    if (el) el.style.display = 'none';
-    return;
-  }
-
-  if (!el) {
-    const actions = document.querySelector('.chat-header-actions');
-    if (!actions) return;
-    el = document.createElement('div');
-    el.className = 'chat-header-usage';
-    el.id = 'chat-header-usage';
-    actions.parentElement.insertBefore(el, actions);
-  }
-
-  // Badge shows session usage (current work)
-  const displayUsage = sessUsage || convUsage;
-  const sessionTokens = (displayUsage.inputTokens || 0) + (displayUsage.outputTokens || 0);
-
-  // Tooltip shows session breakdown + conversation totals
-  let tooltipLines = ['── Session ──'];
-  tooltipLines.push(`Input: ${chatFormatTokenCount(displayUsage.inputTokens)} tokens`);
-  tooltipLines.push(`Output: ${chatFormatTokenCount(displayUsage.outputTokens)} tokens`);
-  const sessCacheTokens = (displayUsage.cacheReadTokens || 0) + (displayUsage.cacheWriteTokens || 0);
-  if (sessCacheTokens > 0) {
-    tooltipLines.push(`Cache read: ${chatFormatTokenCount(displayUsage.cacheReadTokens)}`);
-    tooltipLines.push(`Cache write: ${chatFormatTokenCount(displayUsage.cacheWriteTokens)}`);
-  }
-  if (displayUsage.costUsd > 0) {
-    tooltipLines.push(`Cost: ${chatFormatCost(displayUsage.costUsd)}`);
-  }
-
-  if (convUsage && convUsage !== displayUsage) {
-    const convTotal = (convUsage.inputTokens || 0) + (convUsage.outputTokens || 0);
-    if (convTotal > 0) {
-      tooltipLines.push('');
-      tooltipLines.push('── Conversation ──');
-      tooltipLines.push(`Total: ${chatFormatTokenCount(convTotal)} tokens`);
-      if (convUsage.costUsd > 0) {
-        tooltipLines.push(`Cost: ${chatFormatCost(convUsage.costUsd)}`);
-      }
-    }
-  }
-
-  el.title = tooltipLines.join('\n');
-  el.innerHTML = `<span class="chat-usage-tokens">${chatFormatTokenCount(sessionTokens)} tokens</span>`
-    + (displayUsage.costUsd > 0 ? `<span class="chat-usage-cost">${chatFormatCost(displayUsage.costUsd)}</span>` : '');
-  el.style.display = '';
-}
-
-// ── Context menu ──────────────────────────────────────────────────────────────
-
-function chatShowContextMenu(e, convId) {
-  chatCloseContextMenu();
-  const menu = document.createElement('div');
-  menu.className = 'chat-context-menu';
-  menu.innerHTML = `
-    <button class="chat-context-menu-item" data-action="rename">Rename</button>
-    <button class="chat-context-menu-item danger" data-action="delete">Delete</button>
-  `;
-  menu.style.left = e.clientX + 'px';
-  menu.style.top = e.clientY + 'px';
-
-  menu.querySelectorAll('.chat-context-menu-item').forEach(item => {
-    item.onclick = () => {
-      chatCloseContextMenu();
-      if (item.dataset.action === 'rename') chatRenameConversation(convId);
-      else if (item.dataset.action === 'delete') chatDeleteConversation(convId);
-    };
-  });
-
-  document.body.appendChild(menu);
-  chatContextMenuEl = menu;
-
-  requestAnimationFrame(() => {
-    const rect = menu.getBoundingClientRect();
-    if (rect.right > window.innerWidth) menu.style.left = (window.innerWidth - rect.width - 8) + 'px';
-    if (rect.bottom > window.innerHeight) menu.style.top = (window.innerHeight - rect.height - 8) + 'px';
-  });
-}
-
-function chatCloseContextMenu() {
-  if (chatContextMenuEl) {
-    chatContextMenuEl.remove();
-    chatContextMenuEl = null;
-  }
-}
-
-// ── Message rendering ─────────────────────────────────────────────────────────
-
-function chatRenderMessages() {
-  const container = document.getElementById('chat-messages');
-  if (!container) return;
-
-  if (!chatActiveConv || chatActiveConv.messages.length === 0) {
-    container.innerHTML = `
-      <div class="chat-empty-state">
-        <div class="chat-empty-title">What can I help with?</div>
-        <div class="chat-empty-subtitle">Start a conversation with Agent Cockpit</div>
-        <div class="chat-prompt-cards">
-          <div class="chat-prompt-card" data-prompt="Summarize today's action plan and priorities">Summarize today's action plan and priorities</div>
-          <div class="chat-prompt-card" data-prompt="What are my active tasks and their status?">What are my active tasks and their status?</div>
-          <div class="chat-prompt-card" data-prompt="Review my LinkedIn content strategy">Review my LinkedIn content strategy</div>
-        </div>
-      </div>
-    `;
-    // Wire prompt cards
-    container.querySelectorAll('.chat-prompt-card').forEach(card => {
-      card.onclick = () => {
-        const textarea = document.getElementById('chat-textarea');
-        if (textarea) {
-          textarea.value = card.dataset.prompt;
-          chatAutoResize(textarea);
-          chatSendMessage();
-        }
-      };
-    });
-    return;
-  }
-
-  // Messages are already just the current session (archived sessions live in separate files)
-  const currentSessionMsgs = chatActiveConv.messages;
-
-  let html = '';
-
-  // Session activity overview — aggregate all toolActivity across messages
-  const overviewHtml = chatRenderSessionOverview(currentSessionMsgs);
-  if (overviewHtml) html += overviewHtml;
-
-  for (let mi = 0; mi < currentSessionMsgs.length; mi++) {
-    const msg = currentSessionMsgs[mi];
-
-    const isUser = msg.role === 'user';
-    const backendIcon = !isUser && msg.backend ? getBackendIcon(msg.backend) : null;
-    const avatar = isUser ? '👤' : (backendIcon || DEFAULT_BACKEND_ICON);
-    const avatarClass = !isUser && backendIcon ? ' chat-msg-avatar-svg' : '';
-    const roleLabel = isUser ? 'You' : 'Assistant';
-    const backendLabel = msg.backend ? `<span class="chat-msg-model">${esc(CHAT_BACKENDS.find(b => b.id === msg.backend)?.label || msg.backend)}</span>` : '';
-    const rendered = chatRenderMarkdown(msg.content);
-    const caps = msg.backend ? getBackendCapabilities(msg.backend) : {};
-    const thinkingHtml = msg.thinking && caps.thinking !== false ? chatRenderThinkingBlock(msg.thinking, false) : '';
-    const toolActivityHtml = !isUser && msg.toolActivity ? chatRenderToolActivityBlock(msg.toolActivity) : '';
-
-    // Elapsed time for assistant messages (time since preceding user message)
-    let elapsedLabel = '';
-    if (!isUser && msg.timestamp) {
-      for (let j = mi - 1; j >= 0; j--) {
-        if (currentSessionMsgs[j].role === 'user' && currentSessionMsgs[j].timestamp) {
-          const delta = new Date(msg.timestamp) - new Date(currentSessionMsgs[j].timestamp);
-          if (delta > 0 && delta < 3600000) {
-            elapsedLabel = `<span class="chat-msg-elapsed">${chatFormatElapsed(delta)}</span>`;
-          }
-          break;
-        }
-      }
-    }
-
-    const timeLabel = msg.timestamp ? `<span class="chat-msg-time">${chatFormatTimestamp(msg.timestamp)}${elapsedLabel}</span>` : '';
-
-    html += `
-      <div class="chat-msg ${esc(msg.role)}" data-msg-id="${esc(msg.id)}">
-        <div class="chat-msg-wrapper">
-          <div class="chat-msg-avatar${avatarClass}">${avatar}</div>
-          <div class="chat-msg-body">
-            <div class="chat-msg-role">${roleLabel} ${backendLabel}${timeLabel}</div>
-            <div class="chat-msg-content">${thinkingHtml}${toolActivityHtml}${rendered}</div>
-            <div class="chat-msg-actions">
-              <button class="chat-msg-action" data-action="copy-msg" title="Copy">Copy</button>
-            </div>
-          </div>
-        </div>
-      </div>
-    `;
-  }
-
-  container.innerHTML = html;
-  chatWireMessageActions(container);
-  chatHighlightCode(container);
-
-  // Restore streaming UI if this conversation has an active stream
-  const streamState = chatStreamingState.get(chatActiveConvId);
-  if (streamState) {
-    const msgEl = chatAppendStreamingMessage();
-    streamState.streamingMsgEl = msgEl;
-    chatStartElapsedTimer(chatActiveConvId);
-
-    if (streamState.pendingInteraction) {
-      if (streamState.pendingInteraction.type === 'planApproval') {
-        chatShowPlanApproval(msgEl, chatActiveConvId, streamState.pendingInteraction.planContent);
-      } else if (streamState.pendingInteraction.type === 'userQuestion') {
-        chatShowUserQuestion(msgEl, chatActiveConvId, streamState.pendingInteraction.event);
-      }
-    } else {
-      chatUpdateStreamingContent(msgEl, streamState);
-      if (chatCombinedTools(streamState).length || chatCombinedAgents(streamState).length) {
-        chatStartActivityTimer(chatActiveConvId);
-      }
-    }
-    // else: default typing dots shown by chatAppendStreamingMessage
-  }
-
-  // Render queued messages after all real messages and streaming bubble
-  chatRenderQueuedMessages();
-
-  chatScrollToBottom();
-}
-
-function chatRenderThinkingBlock(thinking, expanded) {
-  const openAttr = expanded ? ' open' : '';
-  return `<details class="chat-thinking-block"${openAttr}>
-    <summary class="chat-thinking-toggle">Thinking</summary>
-    <div class="chat-thinking-content">${chatRenderMarkdown(thinking)}</div>
-  </details>`;
-}
-
-/** Render an outcome badge for a completed tool, e.g. "exit 0", "4 matches", "not found" */
-function chatRenderOutcomeBadge(item) {
-  if (!item.outcome && !item.status) return '';
-  const statusClass = item.status === 'error' ? 'chat-outcome-error'
-    : item.status === 'warning' ? 'chat-outcome-warning'
-    : 'chat-outcome-success';
-  const text = item.outcome ? esc(item.outcome) : '';
-  return text ? `<span class="chat-outcome-badge ${statusClass}">${text}</span>` : '';
-}
-
-/** Render a status indicator (checkmark color varies by outcome status) */
-function chatRenderStatusCheck(item) {
-  if (item.status === 'error') return '<span class="chat-activity-check chat-status-error">\u2717</span>';
-  if (item.status === 'warning') return '<span class="chat-activity-check chat-status-warning">\u2713</span>';
-  return '<span class="chat-activity-check">\u2713</span>';
-}
-
-/** Build a short summary string from a toolActivity array, e.g. "15 ops · 2 agents · 5 read, 2 edited" */
-function chatBuildActivitySummary(toolActivity) {
-  if (!toolActivity || toolActivity.length === 0) return 'No activity';
-  const agents = toolActivity.filter(t => t.isAgent);
-  const tools = toolActivity.filter(t => !t.isAgent);
-  const parts = [];
-  parts.push(`${toolActivity.length} op${toolActivity.length !== 1 ? 's' : ''}`);
-  if (agents.length > 0) {
-    parts.push(`${agents.length} agent${agents.length !== 1 ? 's' : ''}`);
-  }
-  const counts = {};
-  for (const t of tools) { counts[t.tool] = (counts[t.tool] || 0) + 1; }
-  const labels = { Read: 'read', Write: 'written', Edit: 'edited', Bash: 'command', Grep: 'search', Glob: 'glob', WebSearch: 'web search', WebFetch: 'web fetch', TodoWrite: 'task update' };
-  const breakdown = Object.entries(counts).map(([tool, n]) => `${n} ${labels[tool] || tool.toLowerCase()}`);
-  if (breakdown.length > 0) parts.push(breakdown.join(', '));
-  return parts.join(' \u00b7 ');
-}
-
-/** Render a collapsible <details> block for tool activity on completed messages. */
-/** Render a completed agent card (expandable with details on click) */
-function chatRenderAgentCard(t, checkHtml, outcomeBadge, elapsed) {
-  const agentType = esc(t.subagentType || 'agent');
-  const agentDesc = t.description ? escWithCode(t.description) : '';
-  const fullDesc = t.description && t.description.length > 40 ? escWithCode(t.description) : '';
-  const outcomeDetail = t.outcome ? `<span class="chat-agent-detail-outcome">Result: ${esc(t.outcome)}</span>` : '';
-  const hasDetails = fullDesc || outcomeDetail;
-  if (hasDetails) {
-    return `<details class="chat-agent-card chat-agent-card-done chat-agent-expandable">
-      <summary class="chat-agent-card-summary">
-        ${checkHtml}
-        <div class="chat-agent-card-header">
-          <span class="chat-agent-type">${agentType}</span>
-          ${agentDesc ? `<span class="chat-agent-card-desc">${agentDesc}</span>` : ''}
-        </div>
-        ${outcomeBadge}${elapsed ? `<span class="chat-activity-elapsed">${elapsed}</span>` : ''}
-      </summary>
-      <div class="chat-agent-card-details">
-        ${outcomeDetail}
-      </div>
-    </details>`;
-  }
-  return `<div class="chat-agent-card chat-agent-card-done">
-    ${checkHtml}
-    <div class="chat-agent-card-header">
-      <span class="chat-agent-type">${agentType}</span>
-      ${agentDesc ? `<span class="chat-agent-card-desc">${agentDesc}</span>` : ''}
-    </div>
-    ${outcomeBadge}${elapsed ? `<span class="chat-activity-elapsed">${elapsed}</span>` : ''}
-  </div>`;
-}
-
-/** Detect parallel groups among items: consecutive agents with startTimes within threshold */
-const PARALLEL_THRESHOLD_MS = 500;
-function chatGroupParallelItems(items) {
-  const groups = [];
-  let i = 0;
-  while (i < items.length) {
-    // Check if this is the start of a parallel agent group
-    if (items[i].isAgent || items[i]._kind === 'agent') {
-      const groupStart = i;
-      let j = i + 1;
-      while (j < items.length
-        && (items[j].isAgent || items[j]._kind === 'agent')
-        && items[j].startTime && items[groupStart].startTime
-        && Math.abs(items[j].startTime - items[groupStart].startTime) < PARALLEL_THRESHOLD_MS) {
-        j++;
-      }
-      if (j - groupStart > 1) {
-        // Parallel group of agents
-        groups.push({ type: 'parallel', items: items.slice(groupStart, j) });
-        i = j;
-        continue;
-      }
-    }
-    groups.push({ type: 'single', item: items[i] });
-    i++;
-  }
-  return groups;
-}
-
-/** Render a single tool activity item (completed, for persisted display) */
-function chatRenderCompletedItem(t) {
-  const checkHtml = chatRenderStatusCheck(t);
-  const outcomeBadge = chatRenderOutcomeBadge(t);
-  if (t.isAgent || t._kind === 'agent') {
-    const elapsed = t.duration ? chatFormatElapsedShort(t.duration) : '';
-    return chatRenderAgentCard(t, checkHtml, outcomeBadge, elapsed);
-  }
-  const desc = t.description ? escWithCode(t.description) : esc(t.tool || 'Tool');
-  const elapsed = t.duration ? chatFormatElapsedShort(t.duration) : '';
-  return `<div class="chat-activity-history-item">${checkHtml} <span class="chat-activity-history-desc">${desc}</span>${outcomeBadge}${elapsed ? `<span class="chat-activity-elapsed">${elapsed}</span>` : ''}</div>`;
-}
-
-function chatRenderToolActivityBlock(toolActivity) {
-  if (!toolActivity || toolActivity.length === 0) return '';
-  const summary = chatBuildActivitySummary(toolActivity);
-  const groups = chatGroupParallelItems(toolActivity);
-  let itemsHtml = '';
-  for (const group of groups) {
-    if (group.type === 'parallel') {
-      itemsHtml += '<div class="chat-parallel-group">';
-      itemsHtml += '<div class="chat-parallel-label">parallel</div>';
-      for (const t of group.items) {
-        itemsHtml += chatRenderCompletedItem(t);
-      }
-      itemsHtml += '</div>';
-    } else {
-      itemsHtml += chatRenderCompletedItem(group.item);
-    }
-  }
-  return `<details class="chat-activity-block">
-    <summary class="chat-activity-toggle">Activity: ${summary}</summary>
-    <div class="chat-activity-block-content">${itemsHtml}</div>
-  </details>`;
-}
-
-/** Render a session-level activity overview dashboard, aggregating toolActivity from all assistant messages. */
-function chatRenderSessionOverview(messages) {
-  if (!messages || messages.length === 0) return '';
-  const allActivity = [];
-  for (const msg of messages) {
-    if (msg.role === 'assistant' && msg.toolActivity && msg.toolActivity.length > 0) {
-      for (const t of msg.toolActivity) allActivity.push(t);
-    }
-  }
-  if (allActivity.length === 0) return '';
-
-  // Counts
-  const totalOps = allActivity.length;
-  const agents = allActivity.filter(t => t.isAgent);
-  const tools = allActivity.filter(t => !t.isAgent);
-  const totalDuration = allActivity.reduce((sum, t) => sum + (t.duration || 0), 0);
-
-  // Tool type breakdown
-  const toolCounts = {};
-  for (const t of tools) {
-    const name = t.tool || 'Unknown';
-    toolCounts[name] = (toolCounts[name] || 0) + 1;
-  }
-  const sortedTools = Object.entries(toolCounts).sort((a, b) => b[1] - a[1]);
-  const maxCount = sortedTools.length > 0 ? sortedTools[0][1] : 1;
-
-  // Status breakdown
-  let successCount = 0, errorCount = 0, warningCount = 0;
-  for (const t of allActivity) {
-    if (t.status === 'success') successCount++;
-    else if (t.status === 'error') errorCount++;
-    else if (t.status === 'warning') warningCount++;
-  }
-
-  // Build summary line
-  const summaryParts = [`${totalOps} ops`];
-  if (agents.length > 0) summaryParts.push(`${agents.length} agent${agents.length !== 1 ? 's' : ''}`);
-  if (totalDuration > 0) summaryParts.push(chatFormatElapsedShort(totalDuration) + ' total');
-
-  // Build bar chart rows
-  let barsHtml = '';
-  for (const [tool, count] of sortedTools) {
-    const pct = Math.max(4, Math.round((count / maxCount) * 100));
-    barsHtml += `<div class="chat-overview-bar-row">
-      <span class="chat-overview-bar-label">${esc(tool)}</span>
-      <div class="chat-overview-bar-track"><div class="chat-overview-bar-fill" style="width:${pct}%"></div></div>
-      <span class="chat-overview-bar-count">${count}</span>
-    </div>`;
-  }
-
-  // Status pills
-  let statusHtml = '';
-  if (successCount || errorCount || warningCount) {
-    statusHtml = '<div class="chat-overview-status-row">';
-    if (successCount) statusHtml += `<span class="chat-outcome-badge chat-outcome-success">${successCount} success</span>`;
-    if (errorCount) statusHtml += `<span class="chat-outcome-badge chat-outcome-error">${errorCount} error</span>`;
-    if (warningCount) statusHtml += `<span class="chat-outcome-badge chat-outcome-warning">${warningCount} warning</span>`;
-    statusHtml += '</div>';
-  }
-
-  // Agent list
-  let agentsHtml = '';
-  if (agents.length > 0) {
-    agentsHtml = '<div class="chat-overview-agents">';
-    for (const a of agents) {
-      const agentType = esc(a.subagentType || 'agent');
-      const elapsed = a.duration ? chatFormatElapsedShort(a.duration) : '';
-      const desc = a.description ? escWithCode(a.description) : '';
-      const outcomeBadge = chatRenderOutcomeBadge(a);
-      agentsHtml += `<div class="chat-overview-agent-row">
-        <span class="chat-agent-type">${agentType}</span>
-        ${desc ? `<span class="chat-overview-agent-desc">${desc}</span>` : ''}
-        ${outcomeBadge}
-        ${elapsed ? `<span class="chat-activity-elapsed">${elapsed}</span>` : ''}
-      </div>`;
-    }
-    agentsHtml += '</div>';
-  }
-
-  return `<details class="chat-session-overview">
-    <summary class="chat-session-overview-toggle">Session Overview: ${summaryParts.join(' \u00b7 ')}</summary>
-    <div class="chat-session-overview-content">
-      ${statusHtml}
-      ${barsHtml ? `<div class="chat-overview-bars">${barsHtml}</div>` : ''}
-      ${agentsHtml ? `<div class="chat-overview-section"><div class="chat-overview-section-title">Agents</div>${agentsHtml}</div>` : ''}
-    </div>
-  </details>`;
-}
-
-const IMAGE_EXTENSIONS = /\.(png|jpe?g|gif|webp|svg|bmp)$/i;
-
-function chatRenderUploadedFiles(html) {
-  // Replace [Uploaded files: path1, path2] with inline images for image files.
-  // This runs on the final HTML, so the pattern may be inside <p> tags.
-  return html.replace(/<p>\s*\[Uploaded files?:\s*([^\]]+)\]\s*<\/p>|(\[Uploaded files?:\s*([^\]]+)\])/g, (match, pInner, bare, bareInner) => {
-    const pathList = pInner || bareInner;
-    if (!pathList) return match;
-    const paths = pathList.split(',').map(p => p.trim());
-    const parts = [];
-    const nonImages = [];
-    for (const fullPath of paths) {
-      const filename = fullPath.split('/').pop();
-      if (IMAGE_EXTENSIONS.test(filename)) {
-        const segments = fullPath.replace(/\\/g, '/').split('/');
-        const artifactsIdx = segments.lastIndexOf('artifacts');
-        const convId = artifactsIdx >= 0 ? segments[artifactsIdx + 1] : chatActiveConvId;
-        const url = chatApiUrl(`conversations/${encodeURIComponent(convId)}/files/${encodeURIComponent(filename)}`);
-        parts.push(`<div class="chat-inline-image-wrap"><img class="chat-inline-image" src="${url}" alt="${esc(filename)}" title="${esc(filename)}" onclick="chatOpenLightbox(this.src)"></div>`);
-      } else {
-        nonImages.push(filename);
-      }
-    }
-    let result = '';
-    if (nonImages.length) {
-      result += `<p>[Uploaded files: ${nonImages.join(', ')}]</p>`;
-    }
-    if (parts.length) {
-      result += parts.join('');
-    }
-    return result;
-  });
-}
-
-function chatRenderMarkdown(text) {
-  if (!text) return '';
-  if (typeof marked !== 'undefined') {
-    const renderer = new marked.Renderer();
-    const origCode = renderer.code;
-    renderer.code = function(code, language) {
-      let codeText, lang;
-      if (typeof code === 'object' && code !== null) {
-        codeText = code.text || '';
-        lang = code.lang || language || '';
-      } else {
-        codeText = code || '';
-        lang = language || '';
-      }
-      const lineCount = codeText.split('\n').length;
-      const collapsible = lineCount > 50;
-      const langLabel = lang ? esc(lang) : 'code';
-      return `<div class="chat-code-block${collapsible ? ' collapsible collapsed' : ''}">
-        <div class="chat-code-header">
-          <span class="chat-code-lang">${langLabel}</span>
-          <button class="chat-code-copy" onclick="chatCopyCode(this)">Copy</button>
-        </div>
-        <pre><code class="${lang ? 'language-' + esc(lang) : ''}">${esc(codeText)}</code></pre>
-        ${collapsible ? '<div class="chat-code-expand" onclick="chatToggleCodeBlock(this)">Show more</div>' : ''}
-      </div>`;
-    };
-    let html = marked.parse(text, { renderer, breaks: true });
-    return chatRenderUploadedFiles(html);
-  }
-  let html = esc(text).replace(/\n/g, '<br>');
-  return chatRenderUploadedFiles(html);
-}
-
-function chatOpenLightbox(src) {
-  const overlay = document.getElementById('chat-lightbox');
-  const img = document.getElementById('chat-lightbox-img');
-  img.src = src;
-  overlay.classList.add('active');
-  document.addEventListener('keydown', chatLightboxEscHandler);
-}
-
-function chatCloseLightbox() {
-  const overlay = document.getElementById('chat-lightbox');
-  overlay.classList.remove('active');
-  document.removeEventListener('keydown', chatLightboxEscHandler);
-}
-
-function chatLightboxEscHandler(e) {
-  if (e.key === 'Escape') chatCloseLightbox();
-}
-
-function chatHighlightCode(container) {
-  if (typeof hljs === 'undefined') return;
-  container.querySelectorAll('.chat-code-block pre code').forEach(el => {
-    hljs.highlightElement(el);
-  });
-}
-
-function chatCopyCode(btn) {
-  const codeEl = btn.closest('.chat-code-block').querySelector('code');
-  if (codeEl) {
-    navigator.clipboard.writeText(codeEl.textContent).then(() => {
-      btn.textContent = 'Copied!';
-      setTimeout(() => btn.textContent = 'Copy', 1500);
-    });
-  }
-}
-window.chatCopyCode = chatCopyCode;
-
-function chatToggleCodeBlock(expandEl) {
-  const block = expandEl.closest('.chat-code-block');
-  if (block) {
-    block.classList.toggle('collapsed');
-    expandEl.textContent = block.classList.contains('collapsed') ? 'Show more' : 'Show less';
-  }
-}
-window.chatToggleCodeBlock = chatToggleCodeBlock;
-
-function chatWireMessageActions(container) {
-  container.querySelectorAll('.chat-msg-action').forEach(btn => {
-    btn.onclick = () => {
-      const action = btn.dataset.action;
-      const msgEl = btn.closest('.chat-msg');
-      const msgId = msgEl ? msgEl.dataset.msgId : null;
-
-      if (action === 'copy-msg') {
-        const content = msgEl.querySelector('.chat-msg-content');
-        if (content) navigator.clipboard.writeText(content.textContent);
-      }
-    };
-  });
-}
-
-// ── Message Queue UI ─────────────────────────────────────────────────────────
-
-function chatRenderQueuedMessages() {
-  const container = document.getElementById('chat-messages');
-  if (!container) return;
-
-  // Remove any existing queued message elements
-  container.querySelectorAll('.chat-msg-queued').forEach(el => el.remove());
-  container.querySelectorAll('.chat-queue-paused-banner').forEach(el => el.remove());
-
-  const convId = chatActiveConvId;
-  if (!convId) return;
-  const queue = chatMessageQueue.get(convId);
-  if (!queue || queue.length === 0) return;
-
-  // If paused, show a banner before the queued messages
-  if (chatQueuePaused.has(convId)) {
-    const bannerEl = document.createElement('div');
-    bannerEl.className = 'chat-queue-paused-banner';
-    bannerEl.innerHTML = `
-      <span>Queue paused due to error.</span>
-      <button class="chat-queue-resume-btn" onclick="chatResumeQueue()">Resume queue</button>
-      <button class="chat-queue-clear-btn" onclick="chatClearQueue()">Clear queue</button>
-    `;
-    container.appendChild(bannerEl);
-  }
-
-  for (const item of queue) {
-    const el = document.createElement('div');
-    el.className = 'chat-msg user chat-msg-queued' + (item.inFlight ? ' chat-msg-in-flight' : '');
-    el.dataset.queueId = item.id;
-    const rendered = chatRenderMarkdown(item.content);
-    el.innerHTML = `
-      <div class="chat-msg-wrapper">
-        <div class="chat-msg-avatar">👤</div>
-        <div class="chat-msg-body">
-          <div class="chat-msg-role">You <span class="chat-queue-badge">${item.inFlight ? 'Sending...' : 'Queued'}</span></div>
-          <div class="chat-msg-content chat-queued-content">${rendered}</div>
-          ${!item.inFlight ? `<div class="chat-msg-actions chat-queue-actions" style="opacity:1;">
-            <button class="chat-msg-action" data-action="edit-queued" data-queue-id="${item.id}" title="Edit">Edit</button>
-            <button class="chat-msg-action" data-action="delete-queued" data-queue-id="${item.id}" title="Delete">Delete</button>
-          </div>` : ''}
-        </div>
-      </div>
-    `;
-    container.appendChild(el);
-  }
-
-  // Wire queue action buttons
-  container.querySelectorAll('[data-action="delete-queued"]').forEach(btn => {
-    btn.onclick = () => chatDeleteQueuedMessage(Number(btn.dataset.queueId));
-  });
-  container.querySelectorAll('[data-action="edit-queued"]').forEach(btn => {
-    btn.onclick = () => chatEditQueuedMessage(Number(btn.dataset.queueId));
-  });
-}
-
-function chatDeleteQueuedMessage(queueId) {
-  const convId = chatActiveConvId;
-  if (!convId) return;
-  const queue = chatMessageQueue.get(convId);
-  if (!queue) return;
-  const idx = queue.findIndex(item => item.id === queueId);
-  if (idx === -1 || queue[idx].inFlight) return;
-  queue.splice(idx, 1);
-  if (queue.length === 0) {
-    chatMessageQueue.delete(convId);
-    chatQueuePaused.delete(convId);
-  }
-  chatRenderQueuedMessages();
-  chatUpdateSendButtonState();
-}
-
-function chatEditQueuedMessage(queueId) {
-  const convId = chatActiveConvId;
-  if (!convId) return;
-  const queue = chatMessageQueue.get(convId);
-  if (!queue) return;
-  const item = queue.find(i => i.id === queueId);
-  if (!item || item.inFlight) return;
-
-  const container = document.getElementById('chat-messages');
-  if (!container) return;
-  const msgEl = container.querySelector(`.chat-msg-queued[data-queue-id="${queueId}"]`);
-  if (!msgEl) return;
-
-  const contentEl = msgEl.querySelector('.chat-queued-content');
-  const actionsEl = msgEl.querySelector('.chat-queue-actions');
-  if (!contentEl) return;
-
-  // Replace content with textarea for inline editing
-  const editArea = document.createElement('textarea');
-  editArea.className = 'chat-queue-edit-textarea';
-  editArea.value = item.content;
-  editArea.rows = Math.max(2, item.content.split('\n').length);
-
-  const editActions = document.createElement('div');
-  editActions.className = 'chat-queue-edit-actions';
-  editActions.innerHTML = `
-    <button class="chat-queue-edit-save">Save</button>
-    <button class="chat-queue-edit-cancel">Cancel</button>
-  `;
-
-  contentEl.replaceWith(editArea);
-  if (actionsEl) actionsEl.style.display = 'none';
-  editArea.parentElement.appendChild(editActions);
-  editArea.focus();
-  editArea.setSelectionRange(editArea.value.length, editArea.value.length);
-
-  editActions.querySelector('.chat-queue-edit-save').onclick = () => {
-    const newContent = editArea.value.trim();
-    if (newContent) {
-      item.content = newContent;
-    }
-    chatRenderQueuedMessages();
-  };
-
-  editActions.querySelector('.chat-queue-edit-cancel').onclick = () => {
-    chatRenderQueuedMessages();
-  };
-
-  editArea.onkeydown = (e) => {
-    if (e.key === 'Enter' && !e.shiftKey) {
-      e.preventDefault();
-      editActions.querySelector('.chat-queue-edit-save').click();
-    } else if (e.key === 'Escape') {
-      editActions.querySelector('.chat-queue-edit-cancel').click();
-    }
-  };
-}
-
-function chatResumeQueue() {
-  const convId = chatActiveConvId;
-  if (!convId) return;
-  chatQueuePaused.delete(convId);
-  chatRenderQueuedMessages();
-  // If not currently streaming, process the next message
-  if (!chatStreamingConvs.has(convId)) {
-    chatProcessNextQueuedMessage(convId);
-  }
-}
-window.chatResumeQueue = chatResumeQueue;
-
-function chatClearQueue() {
-  const convId = chatActiveConvId;
-  if (!convId) return;
-  const queue = chatMessageQueue.get(convId);
-  if (!queue) return;
-  // Keep in-flight messages, remove the rest
-  const inFlight = queue.filter(i => i.inFlight);
-  if (inFlight.length > 0) {
-    chatMessageQueue.set(convId, inFlight);
-  } else {
-    chatMessageQueue.delete(convId);
-  }
-  chatQueuePaused.delete(convId);
-  chatRenderQueuedMessages();
-  chatUpdateSendButtonState();
-}
-window.chatClearQueue = chatClearQueue;
-
-async function chatProcessNextQueuedMessage(convId) {
-  if (chatQueuePaused.has(convId)) return;
-  if (chatStreamingConvs.has(convId)) return;
-  const queue = chatMessageQueue.get(convId);
-  if (!queue || queue.length === 0) return;
-
-  const nextItem = queue[0];
-  nextItem.inFlight = true;
-  chatRenderQueuedMessages();
-
-  // Populate textarea with the queued content and send
-  const textarea = document.getElementById('chat-textarea');
-  if (textarea) {
-    textarea.value = nextItem.content;
-    chatAutoResize(textarea);
-  }
-
-  // Remove the item from queue before sending (it will become a real message)
-  queue.shift();
-  if (queue.length === 0) chatMessageQueue.delete(convId);
-  chatRenderQueuedMessages();
-
-  // Send the message — chatSendMessage will handle the rest
-  // We need to make sure we're on the right conversation
-  if (chatActiveConvId === convId) {
-    await chatSendMessage();
-  }
-}
-
-function chatScrollToBottom() {
-  const container = document.getElementById('chat-messages');
-  if (container) {
-    requestAnimationFrame(() => {
-      container.scrollTop = container.scrollHeight;
-    });
-  }
-}
-
-// ── Sending messages ──────────────────────────────────────────────────────────
-
-function chatCleanupStreamState(convId, { force = false } = {}) {
-  const st = chatStreamingState.get(convId);
-  if (!st) return;
-  if (st.elapsedTimerInterval) clearInterval(st.elapsedTimerInterval);
-  if (st.activityTimerInterval) clearInterval(st.activityTimerInterval);
-  if (st.pendingInteraction && !force) {
-    // Keep the streaming bubble alive for pending interactions
-    // (plan approval, user questions) so the user can still act on them
-    return;
-  }
-  if (st.streamingMsgEl && st.streamingMsgEl.isConnected) {
-    st.streamingMsgEl.remove();
-  }
-  chatStreamingState.delete(convId);
-}
-
-async function chatSendMessage() {
-  const textarea = document.getElementById('chat-textarea');
-  const hasText = textarea && textarea.value.trim();
-  const completedFiles = chatPendingFiles.filter(e => e.status === 'done');
-  const hasFiles = completedFiles.length > 0;
-  if ((!hasText && !hasFiles) || chatResettingConvs.has(chatActiveConvId)) return;
-  if (chatPendingFiles.some(e => e.status === 'uploading')) return;
-
-  let content = textarea ? textarea.value.trim() : '';
-  if (textarea) { textarea.value = ''; chatAutoResize(textarea); }
-
-  // Gather uploaded file paths and clear pending files
-  if (hasFiles) {
-    const paths = completedFiles.map(e => e.result.path).join(', ');
-    content = content
-      ? content + '\n\n[Uploaded files: ' + paths + ']'
-      : '[Uploaded files: ' + paths + ']';
-  }
-  chatPendingFiles = [];
-  chatRenderFileChips();
-  const sendBtn = document.getElementById('chat-send-btn');
-  if (sendBtn) sendBtn.disabled = true;
-
-  // Create conversation if none active (text-only messages without prior file attach)
-  if (!chatActiveConvId) {
-    chatDraftState.delete('__new__');
-    try {
-      const body = chatPendingWorkingDir ? { workingDir: chatPendingWorkingDir } : {};
-      chatPendingWorkingDir = null;
-      const res = await chatFetch('conversations', { method: 'POST', body });
-      const conv = await res.json();
-      chatActiveConvId = conv.id;
-      chatActiveConv = conv;
-      chatLoadConversations();
-      chatUpdateHeader();
-    } catch (err) {
-      alert('Failed to create conversation: ' + err.message);
-      return;
-    }
-  }
-
-  // If currently streaming, queue the message instead of sending immediately
-  if (chatStreamingConvs.has(chatActiveConvId)) {
-    const queue = chatMessageQueue.get(chatActiveConvId) || [];
-    queue.push({ id: ++chatQueueIdCounter, content, inFlight: false });
-    chatMessageQueue.set(chatActiveConvId, queue);
-    // Un-pause queue if user queues a new message (they've seen the error)
-    chatQueuePaused.delete(chatActiveConvId);
-    chatRenderQueuedMessages();
-    chatUpdateSendButtonState();
-    chatScrollToBottom();
-    return;
-  }
-
-  chatDraftState.delete(chatActiveConvId);
-  const backend = document.getElementById('chat-backend-select')?.value || (CHAT_BACKENDS[0]?.id || 'claude-code');
-  const targetConvId = chatActiveConvId;
-
-  // Start streaming
-  chatStreamingConvs.add(targetConvId);
-  chatStreamingState.set(targetConvId, {
-    assistantContent: '',
-    assistantThinking: '',
-    activeTools: [],
-    activeAgents: [],
-    toolHistory: [],
-    agentHistory: [],
-    planModeActive: false,
-    pendingInteraction: null,
-    streamingMsgEl: null,
-    streamStartTime: Date.now(),
-    elapsedTimerInterval: null,
-    activityTimerInterval: null,
-  });
-  chatRenderConvList();
-  chatUpdateSendButtonState();
-
-  try {
-    // Connect WebSocket before POST so it's ready when the server starts piping events
-    const ws = chatConnectWs(targetConvId);
-    if (ws.readyState !== WebSocket.OPEN) {
-      await new Promise((resolve, reject) => {
-        ws.addEventListener('open', resolve, { once: true });
-        ws.addEventListener('error', reject, { once: true });
-      });
-    }
-
-    const response = await fetch(chatApiUrl(`conversations/${targetConvId}/message`), {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        'x-csrf-token': csrfToken || '',
-      },
-      credentials: 'same-origin',
-      body: JSON.stringify({ content, backend }),
-    });
-
-    if (!response.ok) {
-      const err = await response.json().catch(() => ({}));
-      throw new Error(err.error || response.statusText);
-    }
-
-    const postResult = await response.json();
-
-    if (chatActiveConv && chatActiveConvId === targetConvId && postResult.userMessage) {
-      chatActiveConv.messages.push(postResult.userMessage);
-      chatRenderMessages();
-    }
-
-    const state = chatStreamingState.get(targetConvId);
-    if (!state) return; // cleaned up before stream started
-
-    if (chatActiveConvId === targetConvId && !state.streamingMsgEl) {
-      state.streamingMsgEl = chatAppendStreamingMessage();
-      chatStartElapsedTimer(targetConvId);
-    }
-
-    // Wait for the stream to complete. _doneResolve is called by:
-    // - 'done' event in chatHandleStreamEvent (normal completion)
-    // - chatAutoReconnectWs giving up after max attempts (reconnect exhaustion)
-    // Auto-reconnect on unexpected WS close is handled in chatConnectWs's onclose.
-    await new Promise((resolve) => {
-      const st = chatStreamingState.get(targetConvId);
-      if (st) st._doneResolve = resolve;
-    });
-  } catch (err) {
-    if (err.name !== 'AbortError') {
-      if (chatActiveConvId === targetConvId) chatAppendError(err.message);
-    }
-    // Mark error for queue pause logic
-    const errSt = chatStreamingState.get(targetConvId);
-    if (errSt) errSt._hadError = true;
-  } finally {
-    const finalSt = chatStreamingState.get(targetConvId);
-    const hadError = finalSt?._hadError;
-    chatStreamingConvs.delete(targetConvId);
-    // Force cleanup — stream has ended, no more events will arrive
-    chatCleanupStreamState(targetConvId, { force: true });
-    chatDisconnectWs(targetConvId);
-    chatUpdateSendButtonState();
-    chatRenderConvList();
-    // Refresh conversation list from server to pick up title changes and message counts.
-    // The generation counter in chatLoadConversations discards any stale in-flight responses.
-    chatLoadConversations();
-
-    // Process message queue: pause on error, otherwise send next
-    if (hadError && chatMessageQueue.has(targetConvId)) {
-      chatQueuePaused.add(targetConvId);
-      chatRenderQueuedMessages();
-    } else {
-      chatProcessNextQueuedMessage(targetConvId);
-    }
-  }
-}
-
-/** Handle a single stream event from the WebSocket (or SSE fallback). */
-function chatHandleStreamEvent(targetConvId, event) {
-  const st = chatStreamingState.get(targetConvId);
-  if (!st) return;
-  const isStillActive = (chatActiveConvId === targetConvId);
-
-  // Ensure DOM element exists when active, detect orphaned nodes
-  if (isStillActive && (!st.streamingMsgEl || !st.streamingMsgEl.isConnected)) {
-    st.streamingMsgEl = chatAppendStreamingMessage();
-  }
-
-  if (event.type === 'replay_start') {
-    // Server is replaying buffered events after reconnection — reset streaming state
-    console.log(`[ws] Replay starting: ${event.bufferedEvents} events for conv=${targetConvId}`);
-    st.assistantContent = '';
-    st.assistantThinking = '';
-    st.activeTools = [];
-    st.activeAgents = [];
-    st.toolHistory = [];
-    st.agentHistory = [];
-    st.planModeActive = false;
-    st.pendingInteraction = null;
-    st._replaying = true;
-    // Remove persisted messages that were added during the disconnected stream
-    // (they'll be re-added from replay's assistant_message events)
-    if (isStillActive && chatActiveConv) {
-      chatRenderMessages();
-    }
-    return;
-  } else if (event.type === 'replay_end') {
-    st._replaying = false;
-    chatReconnectAttempts.delete(targetConvId);
-    console.log(`[ws] Replay complete for conv=${targetConvId}`);
-    if (isStillActive) {
-      chatUpdateStreamingContent(st.streamingMsgEl, st);
-    }
-    return;
-  } else if (event.type === 'thinking') {
-    st.assistantThinking += event.content;
-    // Note: do NOT archive active tools here — turn_complete handles that.
-    // Archiving on thinking prematurely kills agent spinners and timers.
-    if (isStillActive) {
-      chatUpdateStreamingContent(st.streamingMsgEl, st);
-    }
-  } else if (event.type === 'tool_outcomes') {
-    // Merge tool outcomes into active tools by matching toolUseId → id
-    for (const outcome of (event.outcomes || [])) {
-      const match = st.activeTools.find(t => t.id === outcome.toolUseId)
-        || st.activeAgents.find(a => a.id === outcome.toolUseId);
-      if (match) {
-        match.outcome = outcome.outcome;
-        match.status = outcome.status;
-      }
-    }
-    if (isStillActive) {
-      chatUpdateStreamingContent(st.streamingMsgEl, st);
-    }
-  } else if (event.type === 'turn_complete') {
-    // Tools finished executing — archive to history so spinners stop but history persists
-    chatArchiveActiveState(st);
-    if (isStillActive && !st.pendingInteraction) {
-      chatUpdateStreamingContent(st.streamingMsgEl, st);
-    }
-  } else if (event.type === 'text') {
-    st.assistantContent += event.content;
-    if (st.pendingInteraction) {
-      // Pending interaction (plan approval, user question) — don't overwrite
-      // dialog with streaming text. Just accumulate assistantContent silently.
-    } else {
-      if (isStillActive) {
-        chatUpdateStreamingContent(st.streamingMsgEl, st);
-      }
-    }
-  } else if (event.type === 'tool_activity') {
-    if (event.isAgent) {
-      st.activeAgents.push({ subagentType: event.subagentType || 'agent', description: event.description || '', startTime: event.startTime || Date.now(), id: event.id, isAgent: true, parentAgentId: event.parentAgentId || null });
-    } else if (event.isPlanMode) {
-      if (event.planAction === 'enter') st.planModeActive = true;
-      else if (event.planAction === 'exit') st.planModeActive = false;
-    }
-    if (!event.isAgent && !event.isPlanMode) {
-      st.activeTools.push({ tool: event.tool, description: event.description || '', startTime: event.startTime || Date.now(), id: event.id, parentAgentId: event.parentAgentId || null });
-    }
-    // Track pending interactions for restoration on switch-back
-    if (event.isPlanMode && event.planAction === 'exit') {
-      // Prefer plan file content (from Write tool) over streamed text summary
-      const planContent = event.planContent || st.assistantContent;
-      st.pendingInteraction = { type: 'planApproval', planContent };
-    } else if (event.isQuestion) {
-      st.pendingInteraction = { type: 'userQuestion', event };
-    }
-    if (isStillActive) {
-      if (event.isPlanMode && event.planAction === 'exit') {
-        chatShowPlanApproval(st.streamingMsgEl, targetConvId, st.pendingInteraction.planContent);
-      } else if (event.isQuestion) {
-        chatShowUserQuestion(st.streamingMsgEl, targetConvId, event);
-      } else if (!st.pendingInteraction) {
-        // Only render tool activity if no pending interaction (plan approval, user question)
-        chatUpdateStreamingContent(st.streamingMsgEl, st);
-        chatStartActivityTimer(targetConvId);
-      }
-    }
-  } else if (event.type === 'assistant_message') {
-    // Reset streaming state before re-render so the restored bubble
-    // shows typing dots instead of stale content duplicating the
-    // completed message that chatRenderMessages is about to display.
-    // Preserve pending interactions (plan approval, user questions)
-    // so they survive intermediate message saves — chatRenderMessages()
-    // will restore the UI from pendingInteraction.
-    const savedInteraction = st.pendingInteraction;
-    st.assistantContent = '';
-    st.assistantThinking = '';
-    chatArchiveActiveState(st);
-    // Clear history — these are now persisted in the saved message's toolActivity.
-    // Keeping them causes duplicates when the next turn adds new agents.
-    st.toolHistory = [];
-    st.agentHistory = [];
-    st.planModeActive = false;
-    st.pendingInteraction = savedInteraction;
-    if (isStillActive && chatActiveConv) {
-      // Avoid duplicate messages during replay
-      if (!chatActiveConv.messages.some(m => m.id === event.message.id)) {
-        chatActiveConv.messages.push(event.message);
-      }
-      chatRenderMessages();
-      chatUpdateHeader();
-    }
-    chatLoadConversations();
-  } else if (event.type === 'title_updated') {
-    // Server generated a new title after a session reset
-    if (isStillActive && chatActiveConv) {
-      chatActiveConv.title = event.title;
-      chatUpdateHeader();
-    }
-    // Update sidebar conversation list
-    const sidebarConv = chatConversations.find(c => c.id === targetConvId);
-    if (sidebarConv) {
-      sidebarConv.title = event.title;
-      chatRenderConvList();
-    }
-  } else if (event.type === 'usage') {
-    // Update usage on active conversation
-    if (isStillActive && chatActiveConv) {
-      chatActiveConv.usage = event.usage;
-      if (event.sessionUsage) chatActiveConv.sessionUsage = event.sessionUsage;
-      chatUpdateUsageDisplay();
-    }
-  } else if (event.type === 'error') {
-    st.pendingInteraction = null;
-    chatArchiveActiveState(st);
-    st.planModeActive = false;
-    st._hadError = true;
-    if (isStillActive) chatAppendError(event.error);
-  } else if (event.type === 'done') {
-    // Ignore stale 'done' during replay — it's from the old stream, not the current one
-    if (st._replaying) return;
-    if (st._doneResolve) { st._doneResolve(); delete st._doneResolve; }
-    chatCleanupStreamState(targetConvId);
-  }
-}
-
-function chatAppendStreamingMessage() {
-  const container = document.getElementById('chat-messages');
-  if (!container) return null;
-
-  const currentBackend = chatActiveConv?.backend || 'claude-code';
-  const icon = getBackendIcon(currentBackend);
-  const msgEl = document.createElement('div');
-  msgEl.className = 'chat-msg assistant streaming';
-  msgEl.innerHTML = `
-    <div class="chat-msg-wrapper">
-      <div class="chat-msg-avatar${icon ? ' chat-msg-avatar-svg' : ''}">${icon || DEFAULT_BACKEND_ICON}</div>
-      <div class="chat-msg-body">
-        <div class="chat-msg-role">Assistant<span class="chat-elapsed-timer"></span></div>
-        <div class="chat-msg-content">
-          <div class="chat-typing">
-            <div class="chat-typing-dot"></div>
-            <div class="chat-typing-dot"></div>
-            <div class="chat-typing-dot"></div>
-          </div>
-        </div>
-      </div>
-    </div>
-  `;
-  container.appendChild(msgEl);
-  chatScrollToBottom();
-  return msgEl;
-}
-
-/** Group sorted items by agent: each agent gets its own group with sub-tools assigned via parentAgentId.
- *  Agents ALWAYS get their own top-level card — even sub-agents with parentAgentId.
- *  Only non-agent tools are nested into their parent agent's sub-activities panel. */
-function chatGroupItemsByAgent(items) {
-  const groups = [];
-  const agentGroupMap = new Map(); // agentId → group object
-  let currentStandalone = null;
-
-  for (const item of items) {
-    if (item._kind === 'agent' || item.isAgent) {
-      // Every agent gets its own card — never nest agents inside other agents
-      currentStandalone = null;
-      const group = { type: 'agent', agent: item, items: [] };
-      groups.push(group);
-      if (item.id) agentGroupMap.set(item.id, group);
-    } else if (item.parentAgentId && agentGroupMap.has(item.parentAgentId)) {
-      // Sub-tool with explicit parent — assign to parent agent's sub-activities
-      currentStandalone = null;
-      agentGroupMap.get(item.parentAgentId).items.push(item);
-    } else {
-      // No parentAgentId — fall back to last agent group (chronological)
-      let lastAgentGroup = null;
-      for (let i = groups.length - 1; i >= 0; i--) {
-        if (groups[i].type === 'agent') { lastAgentGroup = groups[i]; break; }
-      }
-      if (lastAgentGroup) {
-        currentStandalone = null;
-        lastAgentGroup.items.push(item);
-      } else {
-        if (!currentStandalone) {
-          currentStandalone = { type: 'standalone', items: [] };
-          groups.push(currentStandalone);
-        }
-        currentStandalone.items.push(item);
-      }
-    }
-  }
-  return groups;
-}
-
-/** Render a single streaming item (tool or agent) — either running (spinner) or completed (check). */
-function chatRenderStreamingItem(item) {
-  if (item.completed) {
-    return chatRenderCompletedItem(item);
-  }
-  // Running item
-  if (item._kind === 'agent' || item.isAgent) {
-    const agentType = esc(item.subagentType || 'agent');
-    const agentDesc = item.description ? escWithCode(item.description) : '';
-    const initialElapsed = item.startTime ? chatFormatElapsed(Date.now() - item.startTime) : '';
-    return `<div class="chat-agent-card">
-      <div class="chat-agent-spinner" style="animation-delay:-${Date.now() % 800}ms"></div>
-      <div class="chat-agent-card-header">
-        <span class="chat-agent-type">${agentType}</span>
-        ${agentDesc ? `<span class="chat-agent-card-desc">${agentDesc}</span>` : ''}
-      </div>
-      ${initialElapsed ? `<span class="chat-agent-timer-live">${initialElapsed}</span>` : ''}
-    </div>`;
-  }
-  const desc = item.description ? escWithCode(item.description) : esc(item.tool || 'Working');
-  const initialElapsed = item.startTime ? chatFormatElapsed(Date.now() - item.startTime) : '';
-  return `<div class="chat-activity-indicator">
-    <div class="chat-typing"><div class="chat-typing-dot"></div><div class="chat-typing-dot"></div><div class="chat-typing-dot"></div></div>
-    <span class="chat-activity-label">${desc}</span>
-    ${initialElapsed ? `<span class="chat-activity-timer-live">${initialElapsed}</span>` : ''}
-  </div>`;
-}
-
-/** Unified streaming render — stacks text, thinking, tool history, and active tool in one view. */
-function chatUpdateStreamingContent(msgEl, st) {
-  if (!msgEl) return;
-  const contentEl = msgEl.querySelector('.chat-msg-content');
-  if (!contentEl) return;
-
-  let html = '';
-
-  // 1. Thinking block
-  if (st.assistantThinking) {
-    html += chatRenderThinkingBlock(st.assistantThinking, true);
-  }
-
-  // 2. Text content
-  if (st.assistantContent) {
-    html += chatRenderMarkdown(st.assistantContent);
-  } else if (st.assistantThinking) {
-    html += '<div class="chat-thinking-status">Thinking...</div>';
-  }
-
-  // 3. Tool activity — per-agent layout: each agent card followed by its sub-activities
-  const allItems = [
-    ...chatCombinedTools(st).map(t => ({ ...t, _kind: 'tool' })),
-    ...chatCombinedAgents(st).map(a => ({ ...a, _kind: 'agent' })),
-  ].sort((a, b) => (a.startTime || 0) - (b.startTime || 0));
-
-  // Fill in durations for completed items that don't have one yet
-  for (let i = 0; i < allItems.length; i++) {
-    if (allItems[i].completed && !allItems[i].duration && allItems[i].startTime) {
-      const nextStart = allItems[i + 1]?.startTime || Date.now();
-      allItems[i] = { ...allItems[i], duration: nextStart - allItems[i].startTime };
-    }
-  }
-
-  // Group items: standalone tools (before any agent), then agent + following tools
-  const itemGroups = chatGroupItemsByAgent(allItems);
-
-  // Helper: render a single agent card (spinner or checkmark)
-  function renderAgentCard(agent) {
-    const isRunning = !agent.completed;
-    const agentType = esc(agent.subagentType || 'agent');
-    const agentDesc = agent.description ? escWithCode(agent.description) : '';
-    const elapsed = isRunning
-      ? (agent.startTime ? chatFormatElapsed(Date.now() - agent.startTime) : '')
-      : (agent.duration ? chatFormatElapsedShort(agent.duration) : '');
-    const timerClass = isRunning ? 'chat-agent-timer-live' : 'chat-activity-elapsed';
-    return `<div class="chat-agent-card${isRunning ? '' : ' chat-agent-card-done'}">
-      ${isRunning
-        ? `<div class="chat-agent-spinner" style="animation-delay:-${Date.now() % 800}ms"></div>`
-        : chatRenderStatusCheck(agent)}
-      <div class="chat-agent-card-header">
-        <span class="chat-agent-type">${agentType}</span>
-        ${agentDesc ? `<span class="chat-agent-card-desc">${agentDesc}</span>` : ''}
-      </div>
-      ${chatRenderOutcomeBadge(agent)}
-      ${elapsed ? `<span class="${timerClass}">${elapsed}</span>` : ''}
-    </div>`;
-  }
-
-  for (const group of itemGroups) {
-    if (group.type === 'standalone') {
-      for (const item of group.items) {
-        html += chatRenderStreamingItem(item);
-      }
-    } else if (group.type === 'agent') {
-      // Skip completed agents that had no sub-tools and finished near-instantly
-      // (these are planning/delegation agents the model creates before real work)
-      if (group.agent.completed && group.items.length === 0 && (group.agent.duration || 0) < 2000) {
-        continue;
-      }
-      html += renderAgentCard(group.agent);
-      if (group.items.length > 0) {
-        html += '<div class="chat-agent-subactivities">';
-        for (const item of group.items) {
-          html += chatRenderStreamingItem(item);
-        }
-        html += '</div>';
-      }
-    } else if (group.type === 'parallel-agents') {
-      // Legacy path — should not occur with parentAgentId grouping
-      for (const agent of group.agents) {
-        html += renderAgentCard(agent);
-      }
-    }
-  }
-
-  // Post-completion processing indicator: tools/agents finished but model still working
-  const hasCompletedItems = st.toolHistory.length > 0 || st.agentHistory.length > 0;
-  const hasRunningItems = st.activeTools.length > 0 || st.activeAgents.length > 0;
-  if (hasCompletedItems && !hasRunningItems && !st.assistantContent) {
-    html += `<div class="chat-activity-indicator">
-      <div class="chat-typing"><div class="chat-typing-dot"></div><div class="chat-typing-dot"></div><div class="chat-typing-dot"></div></div>
-      <span class="chat-activity-label">Processing...</span>
-    </div>`;
-  }
-
-  // Plan mode banner
-  if (st.planModeActive) {
-    html += `<div class="chat-plan-mode-banner">
-      <span class="chat-plan-mode-icon">📋</span> Planning mode active
-    </div>`;
-  }
-
-  // Fallback: typing dots if nothing to show
-  if (!html) {
-    html = `<div class="chat-activity-indicator">
-      <div class="chat-typing"><div class="chat-typing-dot"></div><div class="chat-typing-dot"></div><div class="chat-typing-dot"></div></div>
-      <span class="chat-activity-label">Working...</span>
-    </div>`;
-  }
-
-  contentEl.innerHTML = html;
-  if (st.assistantContent || st.assistantThinking) chatHighlightCode(contentEl);
-  // Auto-scroll each agent sub-activity panel to show latest item
-  contentEl.querySelectorAll('.chat-agent-subactivities').forEach(el => {
-    el.scrollTop = el.scrollHeight;
-  });
-  chatScrollToBottom();
-}
-
-/** Archive active tools/agents to history before clearing, preserving elapsed durations.
- *  Agents are only archived once they have received their tool_outcomes (outcome/status set).
- *  This prevents sub-tool turn_complete events from prematurely archiving a still-running agent. */
-function chatArchiveActiveState(st) {
-  const now = Date.now();
-  for (const tool of st.activeTools) {
-    st.toolHistory.push({ ...tool, completed: true, duration: tool.startTime ? now - tool.startTime : null });
-  }
-  st.activeTools = [];
-  const stillRunning = [];
-  for (const agent of st.activeAgents) {
-    if (agent.outcome !== undefined || agent.status !== undefined) {
-      st.agentHistory.push({ ...agent, completed: true, duration: agent.startTime ? now - agent.startTime : null });
-    } else {
-      stillRunning.push(agent);
-    }
-  }
-  st.activeAgents = stillRunning;
-  // Only clear timer if no active items remain
-  if (st.activeAgents.length === 0 && st.activeTools.length === 0) {
-    if (st.activityTimerInterval) { clearInterval(st.activityTimerInterval); st.activityTimerInterval = null; }
-  }
-}
-
-/** Build combined tools array (history + active) for rendering. */
-function chatCombinedTools(st) {
-  return [...st.toolHistory, ...st.activeTools];
-}
-
-/** Build combined agents array (history + active) for rendering. */
-function chatCombinedAgents(st) {
-  return [...st.agentHistory, ...st.activeAgents];
-}
-
-function chatStartElapsedTimer(convId) {
-  const state = chatStreamingState.get(convId);
-  if (!state || state.elapsedTimerInterval) return;
-  // Show initial value immediately
-  const timerEl = state.streamingMsgEl?.querySelector('.chat-elapsed-timer');
-  if (timerEl) timerEl.textContent = chatFormatElapsed(Date.now() - state.streamStartTime);
-  state.elapsedTimerInterval = setInterval(() => {
-    const st = chatStreamingState.get(convId);
-    if (!st || !st.streamingMsgEl || !st.streamingMsgEl.isConnected) {
-      clearInterval(state.elapsedTimerInterval);
-      state.elapsedTimerInterval = null;
-      return;
-    }
-    const el = st.streamingMsgEl.querySelector('.chat-elapsed-timer');
-    if (el) el.textContent = chatFormatElapsed(Date.now() - st.streamStartTime);
-  }, 1000);
-}
-
-function chatStartActivityTimer(convId) {
-  const state = chatStreamingState.get(convId);
-  if (!state || state.activityTimerInterval) return;
-  state.activityTimerInterval = setInterval(() => {
-    const st = chatStreamingState.get(convId);
-    if (!st || !st.streamingMsgEl || !st.streamingMsgEl.isConnected) {
-      clearInterval(state.activityTimerInterval);
-      state.activityTimerInterval = null;
-      return;
-    }
-    // Re-render to update all timer values (avoids DOM index mismatch issues)
-    if (st.activeTools.length > 0 || st.activeAgents.length > 0) {
-      chatUpdateStreamingContent(st.streamingMsgEl, st);
-    }
-  }, 1000);
-}
-
-function chatShowPlanApproval(msgEl, convId, planContent) {
-  if (!msgEl) return;
-  const contentEl = msgEl.querySelector('.chat-msg-content');
-  if (!contentEl) return;
-  const planHtml = planContent ? chatRenderMarkdown(planContent) : '';
-  contentEl.innerHTML = `
-    ${planHtml ? `<div class="chat-plan-approval-content">${planHtml}</div>` : ''}
-    <div class="chat-plan-approval">
-      <div class="chat-plan-approval-title">Plan ready for review</div>
-      <div class="chat-plan-approval-actions">
-        <button class="chat-plan-approval-btn approve" data-action="approve">Approve</button>
-        <button class="chat-plan-approval-btn reject" data-action="reject">Reject</button>
-      </div>
-    </div>
-  `;
-  chatHighlightCode(contentEl);
-  contentEl.querySelectorAll('.chat-plan-approval-btn').forEach(btn => {
-    btn.onclick = async () => {
-      const action = btn.dataset.action;
-      const text = action === 'approve' ? 'yes' : 'no';
-      try {
-        chatWsSend(convId, { type: 'input', text });
-        const approvalState = chatStreamingState.get(convId);
-        if (approvalState) {
-          approvalState.pendingInteraction = null;
-          // If the stream has ended (no longer in chatStreamingConvs), fully clean up
-          if (!chatStreamingConvs.has(convId)) {
-            chatCleanupStreamState(convId, { force: true });
-          }
-        }
-        contentEl.innerHTML = `${planHtml ? `<div class="chat-plan-approval-content">${planHtml}</div>` : ''}<div style="font-size:12px;color:var(--muted);font-style:italic;">Plan ${action === 'approve' ? 'approved' : 'rejected'}.</div>`;
-        chatHighlightCode(contentEl);
-      } catch (err) {
-        contentEl.innerHTML = `<div style="font-size:12px;color:var(--danger);">Failed to send response: ${esc(err.message)}</div>`;
-      }
-    };
-  });
-  chatScrollToBottom();
-}
-
-function chatShowUserQuestion(msgEl, convId, event) {
-  if (!msgEl) return;
-  const contentEl = msgEl.querySelector('.chat-msg-content');
-  if (!contentEl) return;
-
-  const questions = event.questions || [];
-  const questionText = questions.length > 0 ? (questions[0].question || 'Input needed') : (event.description || 'Input needed');
-  const options = questions.length > 0 ? (questions[0].options || []) : [];
-
-  let optionsHtml = '';
-  if (options.length > 0) {
-    optionsHtml = '<div class="chat-user-question-options">';
-    for (let i = 0; i < options.length; i++) {
-      const opt = options[i];
-      optionsHtml += `<button class="chat-user-question-option" data-value="${esc(opt.label || opt)}">${esc(opt.label || opt)}${opt.description ? `<br><span style="font-size:11px;color:var(--muted);font-weight:normal;">${esc(opt.description)}</span>` : ''}</button>`;
-    }
-    optionsHtml += '</div>';
-  }
-
-  contentEl.innerHTML = `
-    <div class="chat-user-question">
-      <div class="chat-user-question-header">Question</div>
-      <div class="chat-user-question-text">${escWithCode(questionText)}</div>
-      ${optionsHtml}
-      <input class="chat-user-question-input" type="text" placeholder="Type your answer...">
-      <button class="chat-user-question-submit" disabled>Send</button>
-    </div>
-  `;
-
-  const input = contentEl.querySelector('.chat-user-question-input');
-  const submitBtn = contentEl.querySelector('.chat-user-question-submit');
-
-  // Option buttons
-  contentEl.querySelectorAll('.chat-user-question-option').forEach(optBtn => {
-    optBtn.onclick = () => {
-      contentEl.querySelectorAll('.chat-user-question-option').forEach(b => b.classList.remove('selected'));
-      optBtn.classList.add('selected');
-      input.value = optBtn.dataset.value;
-      submitBtn.disabled = false;
-    };
-  });
-
-  input.oninput = () => { submitBtn.disabled = !input.value.trim(); };
-  input.onkeydown = (e) => { if (e.key === 'Enter' && input.value.trim()) submitBtn.click(); };
-
-  submitBtn.onclick = async () => {
-    const text = input.value.trim();
-    if (!text) return;
-    submitBtn.disabled = true;
-    try {
-      chatWsSend(convId, { type: 'input', text });
-      const questionState = chatStreamingState.get(convId);
-      if (questionState) {
-        questionState.pendingInteraction = null;
-        // If the stream has ended (no longer in chatStreamingConvs), fully clean up
-        if (!chatStreamingConvs.has(convId)) {
-          chatCleanupStreamState(convId, { force: true });
-        }
-      }
-      contentEl.innerHTML = `<div style="font-size:12px;color:var(--muted);font-style:italic;">Answered: ${esc(text)}</div>`;
-    } catch (err) {
-      contentEl.innerHTML = `<div style="font-size:12px;color:var(--danger);">Failed to send response: ${esc(err.message)}</div>`;
-    }
-  };
-
-  chatScrollToBottom();
-}
-
-function chatFormatErrorMessage(errorMsg) {
-  // Parse "API Error: 500 {...json...}" into a friendly message
-  const apiMatch = errorMsg.match(/^API Error:\s*(\d{3})\s*(.*)/s);
-  if (apiMatch) {
-    const code = apiMatch[1];
-    const rest = apiMatch[2].trim();
-    let detail = '';
-    try {
-      const parsed = JSON.parse(rest);
-      detail = (parsed.error && parsed.error.message) || parsed.message || '';
-    } catch {
-      detail = rest;
-    }
-    if (code === '500') return 'The API returned an internal server error. This is usually a temporary issue — try again.';
-    if (code === '429') return 'Rate limit exceeded. Please wait a moment before retrying.';
-    if (code === '529') return 'The API is temporarily overloaded. Please try again shortly.';
-    return `API error ${code}${detail ? ': ' + detail : ''}`;
-  }
-  return errorMsg;
-}
-
-function chatAppendError(errorMsg) {
-  const container = document.getElementById('chat-messages');
-  if (!container) return;
-
-  const friendly = chatFormatErrorMessage(errorMsg);
-  const errEl = document.createElement('div');
-  errEl.className = 'chat-msg assistant';
-  errEl.innerHTML = `
-    <div class="chat-msg-wrapper">
-      <div class="chat-msg-avatar" style="background:#fee2e2;color:#dc2626;">!</div>
-      <div class="chat-msg-body">
-        <div class="chat-msg-role" style="color:#dc2626;">Error</div>
-        <div class="chat-msg-content" style="color:#dc2626;">${esc(friendly)}</div>
-        <div class="chat-msg-actions" style="opacity:1;">
-          <button class="chat-msg-action" onclick="chatRetryLast()">Retry</button>
-        </div>
-      </div>
-    </div>
-  `;
-  container.appendChild(errEl);
-  chatScrollToBottom();
-}
-
-async function chatStopStreaming() {
-  if (!chatActiveConvId) return;
-  chatWsSend(chatActiveConvId, { type: 'abort' });
-}
-
-
-function chatRetryLast() {
-  if (!chatActiveConv) return;
-  const lastUser = [...chatActiveConv.messages].reverse().find(m => m.role === 'user');
-  if (lastUser) {
-    const textarea = document.getElementById('chat-textarea');
-    if (textarea) {
-      textarea.value = lastUser.content;
-      chatAutoResize(textarea);
-      chatSendMessage();
-    }
-  }
-}
-window.chatRetryLast = chatRetryLast;
-
-// ── Session management ────────────────────────────────────────────────────────
-
-async function chatResetSession(convIdOverride) {
-  const convId = typeof convIdOverride === 'string' ? convIdOverride : chatActiveConvId;
-  if (!convId) return;
-  if (chatStreamingConvs.has(convId)) { alert('Cannot reset session while streaming.'); return; }
-  if (chatResettingConvs.has(convId)) return;
-
-  // Enter resetting state
-  chatResettingConvs.add(convId);
-  chatUpdateSendButtonState();
-
-  const resetBtn = document.getElementById('chat-reset-btn');
-  if (resetBtn) { resetBtn.disabled = true; resetBtn.textContent = '↻ Resetting...'; }
-
-  // Show progress indicator in messages area
-  let progressEl = null;
-  if (convId === chatActiveConvId) {
-    const container = document.getElementById('chat-messages');
-    if (container) {
-      progressEl = document.createElement('div');
-      progressEl.className = 'chat-msg assistant';
-      progressEl.id = 'chat-reset-progress';
-      progressEl.innerHTML = `
-        <div class="chat-msg-wrapper">
-          <div class="chat-msg-avatar chat-msg-avatar-svg">${getBackendIcon(chatActiveConv?.backend || 'claude-code')}</div>
-          <div class="chat-msg-body">
-            <div class="chat-msg-role">System</div>
-            <div class="chat-msg-content">
-              <div class="chat-activity-indicator">
-                <div class="chat-typing"><div class="chat-typing-dot"></div><div class="chat-typing-dot"></div><div class="chat-typing-dot"></div></div>
-                <span class="chat-activity-label">Archiving session...</span>
-              </div>
-            </div>
-          </div>
-        </div>`;
-      container.appendChild(progressEl);
-      chatScrollToBottom();
-    }
-  }
-
-  try {
-    const res = await chatFetch(`conversations/${convId}/reset`, { method: 'POST', body: {} });
-    const data = await res.json();
-    if (convId === chatActiveConvId) {
-      chatActiveConv = data.conversation;
-      chatRenderMessages();
-      chatUpdateHeader();
-    }
-    chatLoadConversations();
-  } catch (err) {
-    if (progressEl && progressEl.isConnected) progressEl.remove();
-    alert('Session reset failed: ' + err.message);
-  } finally {
-    chatResettingConvs.delete(convId);
-    const leftover = document.getElementById('chat-reset-progress');
-    if (leftover) leftover.remove();
-    if (resetBtn) { resetBtn.disabled = false; resetBtn.textContent = '↻ Reset'; }
-    chatUpdateSendButtonState();
-  }
-}
-
-async function chatShowSessions() {
-  if (!chatActiveConvId) return;
-  try {
-    const res = await chatFetch(`conversations/${chatActiveConvId}/sessions`);
-    const data = await res.json();
-    const sessions = data.sessions || [];
-
-    let html = '<div class="chat-modal-body">';
-    if (sessions.length === 0) {
-      html += '<div style="color:var(--muted);font-size:13px;">No sessions yet.</div>';
-    } else {
-      for (const s of sessions) {
-        const started = new Date(s.startedAt).toLocaleString();
-        const ended = s.endedAt ? new Date(s.endedAt).toLocaleString() : null;
-        const status = s.isCurrent ? '<span class="pill pill-running">Current</span>' : '';
-        html += `
-          <div style="padding:10px 0;border-bottom:1px solid var(--border);">
-            <div style="display:flex;align-items:center;gap:8px;justify-content:space-between;">
-              <div style="display:flex;align-items:center;gap:8px;">
-                <strong>Session ${s.number}</strong> ${status}
-              </div>
-              <div style="display:flex;gap:6px;">
-                <button class="chat-header-btn chat-dl-session-btn" data-session="${s.number}" style="font-size:11px;padding:2px 10px;cursor:pointer;">Download</button>
-                <button class="chat-header-btn chat-view-session-btn" data-session="${s.number}" style="font-size:11px;padding:2px 10px;cursor:pointer;">View</button>
-              </div>
-            </div>
-            ${s.summary ? `<div style="font-size:12px;color:var(--fg);margin-top:4px;">${esc(s.summary)}</div>` : ''}
-            <div style="font-size:11px;color:var(--muted);margin-top:2px;">
-              Started: ${esc(started)}${ended ? ` — Ended: ${esc(ended)}` : ''}
-              · ${s.messageCount} messages
-            </div>
-          </div>
-        `;
-      }
-    }
-    html += '</div>';
-
-    chatShowModal('Session History', html);
-    document.querySelectorAll('.chat-view-session-btn').forEach(btn => {
-      btn.addEventListener('click', () => chatViewSession(Number(btn.dataset.session)));
-    });
-    document.querySelectorAll('.chat-dl-session-btn').forEach(btn => {
-      btn.addEventListener('click', () => {
-        window.open(apiUrl(`chat/conversations/${chatActiveConvId}/sessions/${btn.dataset.session}/download`), '_blank');
-      });
-    });
-  } catch (err) {
-    alert('Failed to load sessions: ' + err.message);
-  }
-}
-
-async function chatViewSession(sessionNumber) {
-  if (!chatActiveConv) return;
-  chatCloseModal();
-
-  let sessionMsgs;
-  try {
-    if (sessionNumber === chatActiveConv.sessionNumber) {
-      sessionMsgs = chatActiveConv.messages;
-    } else {
-      const res = await chatFetch(`conversations/${chatActiveConvId}/sessions/${sessionNumber}/messages`);
-      const data = await res.json();
-      sessionMsgs = data.messages || [];
-    }
-  } catch (err) {
-    alert('Failed to load session: ' + err.message);
-    return;
-  }
-
-  let sessionDate = '';
-  if (sessionMsgs.length > 0 && sessionMsgs[0].timestamp) {
-    sessionDate = new Date(sessionMsgs[0].timestamp).toLocaleDateString('en-US', { month: 'short', day: 'numeric', year: 'numeric' });
-  }
-
-  let msgsHtml = '';
-  if (sessionMsgs.length === 0) {
-    msgsHtml = '<div style="color:var(--muted);font-size:13px;padding:16px 0;">No messages in this session.</div>';
-  } else {
-    for (const msg of sessionMsgs) {
-      const isUser = msg.role === 'user';
-      const backendIcon = !isUser && msg.backend ? getBackendIcon(msg.backend) : null;
-      const avatar = isUser ? '👤' : (backendIcon || DEFAULT_BACKEND_ICON);
-      const avatarClass = !isUser && backendIcon ? ' chat-msg-avatar-svg' : '';
-      const roleLabel = isUser ? 'You' : 'Assistant';
-      const backendLabel = msg.backend ? `<span class="chat-msg-model">${esc(CHAT_BACKENDS.find(b => b.id === msg.backend)?.label || msg.backend)}</span>` : '';
-      const rendered = chatRenderMarkdown(msg.content);
-      const caps = msg.backend ? getBackendCapabilities(msg.backend) : {};
-      const thinkingHtml = msg.thinking && caps.thinking !== false ? chatRenderThinkingBlock(msg.thinking, false) : '';
-      const toolActivityHtml = !isUser && msg.toolActivity ? chatRenderToolActivityBlock(msg.toolActivity) : '';
-      msgsHtml += `
-        <div class="chat-msg ${esc(msg.role)}">
-          <div class="chat-msg-wrapper">
-            <div class="chat-msg-avatar${avatarClass}">${avatar}</div>
-            <div class="chat-msg-body">
-              <div class="chat-msg-role">${roleLabel} ${backendLabel}</div>
-              <div class="chat-msg-content">${thinkingHtml}${toolActivityHtml}${rendered}</div>
-            </div>
-          </div>
-        </div>
-      `;
-    }
-  }
-
-  const title = `Session ${sessionNumber}` + (sessionDate ? ` — ${sessionDate}` : '');
-  const html = `
-    <div class="chat-modal-body" style="padding:0;">
-      <div style="padding:8px 16px;border-bottom:1px solid var(--border);">
-        <button class="chat-header-btn" id="chat-back-to-sessions" style="font-size:12px;cursor:pointer;">← Back to Session List</button>
-      </div>
-      <div style="max-height:60vh;overflow-y:auto;padding:8px 16px;">
-        ${msgsHtml}
-      </div>
-    </div>
-  `;
-
-  chatShowModal(title, html);
-
-  document.getElementById('chat-back-to-sessions')?.addEventListener('click', chatShowSessions);
-
-  const overlay = document.getElementById('chat-modal-overlay');
-  if (overlay) chatHighlightCode(overlay);
-}
-window.chatViewSession = chatViewSession;
-
-// ── Download ──────────────────────────────────────────────────────────────────
-
-async function chatDownloadConversation() {
-  if (!chatActiveConvId) return;
-  try {
-    const res = await fetch(chatApiUrl(`conversations/${chatActiveConvId}/download`), { credentials: 'same-origin' });
-    const blob = await res.blob();
-    const disposition = res.headers.get('Content-Disposition') || '';
-    const match = disposition.match(/filename="(.+?)"/);
-    const filename = match ? match[1] : 'conversation.md';
-    const url = URL.createObjectURL(blob);
-    const a = document.createElement('a');
-    a.href = url;
-    a.download = filename;
-    a.click();
-    URL.revokeObjectURL(url);
-  } catch (err) {
-    alert('Download failed: ' + err.message);
-  }
-}
-
-// ── Settings ──────────────────────────────────────────────────────────────────
-
-async function chatShowSettings(initialTab) {
-  try {
-    const res = await chatFetch('settings');
-    chatSettingsData = await res.json();
-  } catch {
-    chatSettingsData = { theme: 'system', sendBehavior: 'enter', systemPrompt: '', defaultBackend: CHAT_BACKENDS[0]?.id || 'claude-code' };
-  }
-
-  const s = chatSettingsData;
-  const html = `
-    <div class="chat-settings-tabs">
-      <button class="chat-settings-tab active" data-tab="general">General</button>
-      <button class="chat-settings-tab" data-tab="usage">Usage Stats</button>
-    </div>
-    <div class="chat-modal-body">
-      <div class="chat-settings-tab-content" id="chat-tab-general">
-        <div class="chat-settings-group">
-          <div class="chat-settings-label">Theme</div>
-          <select class="chat-settings-select" id="chat-settings-theme">
-            <option value="light"${s.theme === 'light' ? ' selected' : ''}>Light</option>
-            <option value="dark"${s.theme === 'dark' ? ' selected' : ''}>Dark</option>
-            <option value="system"${s.theme === 'system' ? ' selected' : ''}>System</option>
-          </select>
-        </div>
-        <div class="chat-settings-group">
-          <div class="chat-settings-label">Send Behavior</div>
-          <select class="chat-settings-select" id="chat-settings-send">
-            <option value="enter"${s.sendBehavior === 'enter' ? ' selected' : ''}>Enter to send</option>
-            <option value="ctrl-enter"${s.sendBehavior === 'ctrl-enter' ? ' selected' : ''}>Ctrl+Enter to send</option>
-          </select>
-        </div>
-        <div class="chat-settings-group">
-          <div class="chat-settings-label">Default Backend</div>
-          <select class="chat-settings-select" id="chat-settings-backend">
-            ${CHAT_BACKENDS.map(b => `<option value="${b.id}"${s.defaultBackend === b.id ? ' selected' : ''}>${esc(b.label)}</option>`).join('')}
-          </select>
-        </div>
-        <div class="chat-settings-group">
-          <div class="chat-settings-label">System Prompt</div>
-          <div class="chat-settings-desc">Prepended to every new CLI session.</div>
-          <textarea class="chat-settings-textarea" id="chat-settings-system-prompt" style="min-height:120px">${esc(s.systemPrompt || '')}</textarea>
-        </div>
-        <button class="chat-settings-save" onclick="chatSaveSettings()">Save Settings</button>
-      </div>
-      <div class="chat-settings-tab-content" id="chat-tab-usage" style="display:none;">
-        <div class="chat-usage-stats-controls">
-          <div class="chat-settings-group" style="flex:1;">
-            <div class="chat-settings-label">Time Range</div>
-            <select class="chat-settings-select" id="chat-usage-range" onchange="chatUpdateUsageStats()">
-              <option value="today">Today</option>
-              <option value="week" selected>This Week</option>
-              <option value="month">This Month</option>
-              <option value="all">All Time</option>
-            </select>
-          </div>
-          <button class="chat-settings-save chat-usage-clear-btn" onclick="chatClearUsageStats()">Clear All Data</button>
-        </div>
-        <div id="chat-usage-stats-body">
-          <div class="chat-usage-loading">Loading...</div>
-        </div>
-      </div>
-    </div>
-  `;
-
-  chatShowModal('Settings', html);
-
-  // Wire up tabs
-  document.querySelectorAll('.chat-settings-tab').forEach(tab => {
-    tab.addEventListener('click', () => {
-      document.querySelectorAll('.chat-settings-tab').forEach(t => t.classList.remove('active'));
-      tab.classList.add('active');
-      document.querySelectorAll('.chat-settings-tab-content').forEach(c => c.style.display = 'none');
-      const target = tab.getAttribute('data-tab');
-      const panel = document.getElementById('chat-tab-' + target);
-      if (panel) panel.style.display = '';
-      if (target === 'usage') chatLoadUsageStats();
-    });
-  });
-
-  // If opening to usage tab
-  if (initialTab === 'usage') {
-    document.querySelector('.chat-settings-tab[data-tab="usage"]')?.click();
-  }
-}
-
-async function chatSaveSettings() {
-  const settings = {
-    theme: document.getElementById('chat-settings-theme')?.value || 'system',
-    sendBehavior: document.getElementById('chat-settings-send')?.value || 'enter',
-    defaultBackend: document.getElementById('chat-settings-backend')?.value || (CHAT_BACKENDS[0]?.id || 'claude-code'),
-    systemPrompt: document.getElementById('chat-settings-system-prompt')?.value || '',
-  };
-  applyTheme(settings.theme);
-  try {
-    await chatFetch('settings', { method: 'PUT', body: settings });
-    chatSettingsData = settings;
-    chatCloseModal();
-  } catch (err) {
-    alert('Failed to save settings: ' + err.message);
-  }
-}
-window.chatSaveSettings = chatSaveSettings;
-
-// ── Usage Stats ──────────────────────────────────────────────────────────────
-
-let _usageStatsCache = null;
-
-async function chatLoadUsageStats() {
-  try {
-    const res = await chatFetch('usage-stats');
-    _usageStatsCache = await res.json();
-    chatUpdateUsageStats();
-  } catch (err) {
-    const body = document.getElementById('chat-usage-stats-body');
-    if (body) body.innerHTML = '<div class="chat-usage-loading">Failed to load usage stats.</div>';
-  }
-}
-
-function chatUpdateUsageStats() {
-  const body = document.getElementById('chat-usage-stats-body');
-  if (!body || !_usageStatsCache) return;
-
-  const range = document.getElementById('chat-usage-range')?.value || 'week';
-  const days = _usageStatsCache.days || [];
-
-  // Filter days by range
-  const now = new Date();
-  const todayStr = now.toISOString().slice(0, 10);
-  let filteredDays;
-  if (range === 'today') {
-    filteredDays = days.filter(d => d.date === todayStr);
-  } else if (range === 'week') {
-    const weekAgo = new Date(now);
-    weekAgo.setDate(weekAgo.getDate() - 7);
-    const weekStr = weekAgo.toISOString().slice(0, 10);
-    filteredDays = days.filter(d => d.date >= weekStr);
-  } else if (range === 'month') {
-    const monthAgo = new Date(now);
-    monthAgo.setDate(monthAgo.getDate() - 30);
-    const monthStr = monthAgo.toISOString().slice(0, 10);
-    filteredDays = days.filter(d => d.date >= monthStr);
-  } else {
-    filteredDays = days;
-  }
-
-  // Normalize: support both old format (backends map) and new format (records array)
-  function getDayRecords(day) {
-    if (day.records) return day.records;
-    // Migrate old format
-    if (day.backends) {
-      return Object.entries(day.backends).map(([bid, u]) => ({ backend: bid, model: 'unknown', usage: u }));
-    }
-    return [];
-  }
-
-  // Aggregate per backend+model
-  const aggregateKey = (r) => `${r.backend}\0${r.model}`;
-  const totals = {};
-  for (const day of filteredDays) {
-    for (const rec of getDayRecords(day)) {
-      const key = aggregateKey(rec);
-      if (!totals[key]) totals[key] = { backend: rec.backend, model: rec.model, usage: { inputTokens: 0, outputTokens: 0, cacheReadTokens: 0, cacheWriteTokens: 0, costUsd: 0 } };
-      totals[key].usage.inputTokens += rec.usage.inputTokens || 0;
-      totals[key].usage.outputTokens += rec.usage.outputTokens || 0;
-      totals[key].usage.cacheReadTokens += rec.usage.cacheReadTokens || 0;
-      totals[key].usage.cacheWriteTokens += rec.usage.cacheWriteTokens || 0;
-      totals[key].usage.costUsd += rec.usage.costUsd || 0;
-    }
-  }
-
-  const keys = Object.keys(totals);
-  if (keys.length === 0) {
-    body.innerHTML = '<div class="chat-usage-loading">No usage data for this period.</div>';
-    return;
-  }
-
-  function backendLabel(id) {
-    const b = CHAT_BACKENDS.find(x => x.id === id);
-    return b ? b.label : id;
-  }
-
-  function modelLabel(m) {
-    if (!m || m === 'unknown') return '-';
-    return m;
-  }
-
-  let html = '<table class="chat-usage-table"><thead><tr><th>Backend</th><th>Model</th><th>Input</th><th>Output</th><th>Cache R</th><th>Cache W</th><th>Total</th><th>Cost</th></tr></thead><tbody>';
-  for (const key of keys) {
-    const t = totals[key];
-    const u = t.usage;
-    const total = u.inputTokens + u.outputTokens;
-    html += `<tr>
-      <td>${esc(backendLabel(t.backend))}</td>
-      <td>${esc(modelLabel(t.model))}</td>
-      <td>${chatFormatTokenCount(u.inputTokens)}</td>
-      <td>${chatFormatTokenCount(u.outputTokens)}</td>
-      <td>${chatFormatTokenCount(u.cacheReadTokens)}</td>
-      <td>${chatFormatTokenCount(u.cacheWriteTokens)}</td>
-      <td>${chatFormatTokenCount(total)}</td>
-      <td>${u.costUsd > 0 ? chatFormatCost(u.costUsd) : '-'}</td>
-    </tr>`;
-  }
-  html += '</tbody></table>';
-
-  // Daily breakdown
-  if (filteredDays.length > 1) {
-    html += '<div class="chat-usage-daily-title">Daily Breakdown</div>';
-    html += '<table class="chat-usage-table chat-usage-daily"><thead><tr><th>Date</th><th>Backend</th><th>Model</th><th>Tokens</th><th>Cost</th></tr></thead><tbody>';
-    const sortedDays = [...filteredDays].sort((a, b) => b.date.localeCompare(a.date));
-    for (const day of sortedDays) {
-      for (const rec of getDayRecords(day)) {
-        const total = (rec.usage.inputTokens || 0) + (rec.usage.outputTokens || 0);
-        html += `<tr>
-          <td>${day.date}</td>
-          <td>${esc(backendLabel(rec.backend))}</td>
-          <td>${esc(modelLabel(rec.model))}</td>
-          <td>${chatFormatTokenCount(total)}</td>
-          <td>${rec.usage.costUsd > 0 ? chatFormatCost(rec.usage.costUsd) : '-'}</td>
-        </tr>`;
-      }
-    }
-    html += '</tbody></table>';
-  }
-
-  body.innerHTML = html;
-}
-window.chatUpdateUsageStats = chatUpdateUsageStats;
-
-async function chatClearUsageStats() {
-  if (!confirm('Clear all usage statistics? This cannot be undone.')) return;
-  try {
-    await chatFetch('usage-stats', { method: 'DELETE' });
-    _usageStatsCache = { days: [] };
-    chatUpdateUsageStats();
-  } catch (err) {
-    alert('Failed to clear usage stats: ' + err.message);
-  }
-}
-window.chatClearUsageStats = chatClearUsageStats;
-
-// ── Workspace Instructions modal ─────────────────────────────────────────────
-
-async function chatShowWorkspaceInstructions(hash, label) {
-  let instructions = '';
-  try {
-    const res = await chatFetch(`workspaces/${encodeURIComponent(hash)}/instructions`);
-    const data = await res.json();
-    instructions = data.instructions || '';
-  } catch {
-    // Workspace may not have instructions yet
-  }
-
-  const html = `
-    <div class="chat-modal-body">
-      <div class="chat-settings-group">
-        <div class="chat-settings-desc">Additional instructions prepended to every new CLI session in this workspace. Combined with the global system prompt.</div>
-        <textarea class="chat-settings-textarea" id="chat-ws-instructions" style="min-height:160px">${esc(instructions)}</textarea>
-      </div>
-      <button class="chat-settings-save" onclick="chatSaveWorkspaceInstructions('${esc(hash)}')">Save</button>
-    </div>
-  `;
-
-  chatShowModal(`Instructions: ${label}`, html);
-}
-
-async function chatSaveWorkspaceInstructions(hash) {
-  const instructions = document.getElementById('chat-ws-instructions')?.value || '';
-  try {
-    await chatFetch(`workspaces/${encodeURIComponent(hash)}/instructions`, {
-      method: 'PUT',
-      body: { instructions },
-    });
-    chatCloseModal();
-  } catch (err) {
-    alert('Failed to save workspace instructions: ' + err.message);
-  }
-}
-window.chatSaveWorkspaceInstructions = chatSaveWorkspaceInstructions;
-
-// ── Modal helper ──────────────────────────────────────────────────────────────
-
-function chatShowModal(title, bodyHtml) {
-  chatCloseModal();
-  const overlay = document.createElement('div');
-  overlay.className = 'chat-modal-overlay';
-  overlay.id = 'chat-modal-overlay';
-  overlay.innerHTML = `
-    <div class="chat-modal">
-      <div class="chat-modal-header">
-        <div class="chat-modal-title">${esc(title)}</div>
-        <button class="chat-modal-close" id="chat-modal-close-btn">✕</button>
-      </div>
-      ${bodyHtml}
-    </div>
-  `;
-  overlay.addEventListener('click', (e) => { if (e.target === overlay) chatCloseModal(); });
-  document.body.appendChild(overlay);
-  document.getElementById('chat-modal-close-btn').addEventListener('click', chatCloseModal);
-}
-
-function chatCloseModal() {
-  const overlay = document.getElementById('chat-modal-overlay');
-  if (overlay) overlay.remove();
-}
-window.chatCloseModal = chatCloseModal;
-
-// ── Self-update UI ───────────────────────────────────────────────────────────
-
-async function chatManualVersionCheck() {
-  const textEl = document.getElementById('chat-version-text');
-  if (!textEl) return;
-  const original = textEl.textContent;
-  textEl.textContent = 'checking…';
-  try {
-    const res = await chatFetch('check-version', { method: 'POST' });
-    const status = await res.json();
-    textEl.textContent = 'v' + (status.localVersion || original.replace(/^v/, ''));
-    chatCheckUpdateIndicator(status);
-  } catch {
-    textEl.textContent = original;
-  }
-}
-
-function chatCheckUpdateIndicator(status) {
-  const indicator = document.getElementById('chat-update-indicator');
-  if (!indicator) return;
-  if (status.updateAvailable && status.remoteVersion) {
-    indicator.textContent = 'v' + status.remoteVersion + ' available';
-    indicator.title = 'Update to v' + status.remoteVersion;
-    indicator.style.display = '';
-    indicator.onclick = () => chatShowUpdateModal(status);
-  } else {
-    indicator.style.display = 'none';
-    indicator.onclick = null;
-  }
-}
-
-function chatShowUpdateModal(status) {
-  const localVer = status.localVersion || status.version || '?';
-  const remoteVer = status.remoteVersion || '?';
-  const html = `
-    <div class="chat-modal-body" style="padding:16px;">
-      <div style="margin-bottom:16px;">
-        <div style="font-size:13px;color:var(--muted);margin-bottom:6px;">
-          Current version: <strong>v${esc(localVer)}</strong>
-        </div>
-        <div style="font-size:13px;color:var(--muted);margin-bottom:16px;">
-          Available version: <strong>v${esc(remoteVer)}</strong>
-        </div>
-        <div style="font-size:12px;color:var(--muted);margin-bottom:16px;">
-          This will pull the latest code from main, install dependencies, and restart the server.
-          The page will reload automatically.
-        </div>
-      </div>
-      <div id="chat-update-status" style="display:none;margin-bottom:12px;"></div>
-      <button class="chat-settings-save" id="chat-update-confirm-btn">Update Now</button>
-    </div>
-  `;
-  chatShowModal('Update Available', html);
-  document.getElementById('chat-update-confirm-btn').addEventListener('click', chatTriggerUpdate);
-}
-
-async function chatTriggerUpdate() {
-  const statusEl = document.getElementById('chat-update-status');
-  const btn = document.getElementById('chat-update-confirm-btn');
-  if (statusEl) {
-    statusEl.style.display = 'block';
-    statusEl.innerHTML = '<div style="color:var(--muted);font-size:13px;">Updating... This may take a moment.</div>';
-  }
-  if (btn) {
-    btn.disabled = true;
-    btn.textContent = 'Updating...';
-  }
-
-  try {
-    const res = await chatFetch('update-trigger', { method: 'POST' });
-    const result = await res.json();
-
-    if (result.success) {
-      if (statusEl) {
-        statusEl.innerHTML = '<div style="color:var(--done);font-size:13px;">Update successful! Restarting server...</div>';
-      }
-      setTimeout(() => chatShowRestartOverlay(), 1000);
-      setTimeout(() => window.location.reload(), 6000);
-    } else {
-      if (statusEl) {
-        const stepsHtml = (result.steps || []).map(s =>
-          '<div style="font-size:12px;color:' + (s.success ? 'var(--done)' : 'var(--danger)') + ';">'
-          + (s.success ? '&#10003; ' : '&#10007; ') + esc(s.name) + '</div>'
-        ).join('');
-        statusEl.innerHTML =
-          '<div style="color:var(--danger);font-size:13px;margin-bottom:8px;">'
-          + esc(result.error) + '</div>' + stepsHtml;
-      }
-      if (btn) {
-        btn.disabled = false;
-        btn.textContent = 'Retry Update';
-      }
-    }
-  } catch (err) {
-    // If the server restarted before sending the response, we get a network error
-    if (err.message === 'Failed to fetch' || err.name === 'TypeError') {
-      chatShowRestartOverlay();
-      setTimeout(() => window.location.reload(), 5000);
-      return;
-    }
-    if (statusEl) {
-      statusEl.innerHTML = '<div style="color:var(--danger);font-size:13px;">Update failed: ' + esc(err.message) + '</div>';
-    }
-    if (btn) {
-      btn.disabled = false;
-      btn.textContent = 'Retry Update';
-    }
-  }
-}
-
-function chatShowRestartOverlay() {
-  chatCloseModal();
-  const overlay = document.createElement('div');
-  overlay.id = 'chat-restart-overlay';
-  overlay.innerHTML =
-    '<div class="restart-dialog">'
-    + '<div style="font-size:18px;font-weight:600;margin-bottom:8px;">Restarting Server...</div>'
-    + '<div style="font-size:13px;color:var(--muted);">The page will reload automatically.</div>'
-    + '</div>';
-  document.body.appendChild(overlay);
-}
-
-// ── Keyboard shortcuts ────────────────────────────────────────────────────────
-
-document.addEventListener('keydown', (e) => {
-  const mod = e.metaKey || e.ctrlKey;
-
-  // Ctrl/Cmd + Shift + O — New chat
-  if (mod && e.shiftKey && e.key === 'O') {
-    e.preventDefault();
-    chatNewConversation();
-    return;
-  }
-
-  // Ctrl/Cmd + K — Search conversations
-  if (mod && e.key === 'k') {
-    e.preventDefault();
-    const searchInput = document.getElementById('chat-search-input');
-    if (searchInput) searchInput.focus();
-    return;
-  }
-
-  // Ctrl/Cmd + Shift + S — Toggle sidebar
-  if (mod && e.shiftKey && e.key === 'S') {
-    e.preventDefault();
-    chatToggleSidebar();
-    return;
-  }
-
-  // Ctrl/Cmd + Shift + C — Copy last response
-  if (mod && e.shiftKey && e.key === 'C') {
-    e.preventDefault();
-    if (chatActiveConv) {
-      const lastAssistant = [...chatActiveConv.messages].reverse().find(m => m.role === 'assistant');
-      if (lastAssistant) navigator.clipboard.writeText(lastAssistant.content);
-    }
-    return;
-  }
-
-  // Ctrl/Cmd + Shift + R — Reset session
-  if (mod && e.shiftKey && e.key === 'R') {
-    e.preventDefault();
-    chatResetSession();
-    return;
-  }
-
-  // Ctrl/Cmd + Shift + D — Download conversation
-  if (mod && e.shiftKey && e.key === 'D') {
-    e.preventDefault();
-    chatDownloadConversation();
-    return;
-  }
-
-  // / — Focus input (when not in input)
-  if (e.key === '/' && !e.target.closest('input, textarea, select, [contenteditable]')) {
-    e.preventDefault();
-    const textarea = document.getElementById('chat-textarea');
-    if (textarea) textarea.focus();
-    return;
-  }
-});
-
-// ── Init ──────────────────────────────────────────────────────────────────────
-chatInit();
-
