+++ public/index.html
@@ -78,12 +78,12 @@
 </div>
 
 <!-- Lightbox overlay for full-size images -->
-<div class="chat-lightbox" id="chat-lightbox" onclick="chatCloseLightbox()">
-  <img class="chat-lightbox-img" id="chat-lightbox-img" src="" alt="" onclick="event.stopPropagation()">
+<div class="chat-lightbox" id="chat-lightbox">
+  <img class="chat-lightbox-img" id="chat-lightbox-img" src="" alt="">
 </div>
 
 <script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.9.0/highlight.min.js"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/marked/12.0.2/marked.min.js"></script>
-<script src="app.js?v=4"></script>
+<script type="module" src="js/main.js"></script>
 </body>
 </html>
